+++ .github/workflows/run-tests-on-push.yml
@@ -6,6 +6,20 @@ env:
   LOG_CONFIG: test
 
 jobs:
+  run-tests-3_9-core-dependencies:
+    runs-on: ubuntu-20.04
+    name: Pytest on Python 3.9 / Ubuntu 20.04
+    steps:
+    - uses: actions/checkout@v4
+    - uses: actions/setup-python@v5
+      with:
+        python-version: "3.9"
+        cache: 'pip'
+    - run: pip install --upgrade pip
+    - run: pip install poetry
+    - run: poetry install --with dev
+    - run: poetry run pytest tests/
+
   run-tests-3_9:
     runs-on: ubuntu-20.04
     name: Pytest on Python 3.9 / Ubuntu 20.04
@@ -20,6 +34,20 @@ jobs:
     - run: poetry install --with dev --extras server
     - run: poetry run pytest tests/ --show-capture=stdout --cov=src
 
+  run-tests-3_10-core-dependencies:
+    runs-on: ubuntu-latest
+    name: Pytest on Python 3.10
+    steps:
+    - uses: actions/checkout@v4
+    - uses: actions/setup-python@v5
+      with:
+        python-version: "3.10"
+        cache: 'pip'
+    - run: pip install --upgrade pip
+    - run: pip install poetry
+    - run: poetry install --with dev
+    - run: poetry run pytest tests/
+
   run-tests-3_10:
     runs-on: ubuntu-latest
     name: Pytest on Python 3.10
@@ -33,6 +61,20 @@ jobs:
     - run: pip install poetry
     - run: poetry install --with dev --extras server
     - run: poetry run pytest tests/ --show-capture=stdout --cov=src
+
+  run-tests-3_11-core-dependencies:
+    runs-on: ubuntu-latest
+    name: Pytest on Python 3.11
+    steps:
+    - uses: actions/checkout@v4
+    - uses: actions/setup-python@v5
+      with:
+        python-version: "3.11"
+        cache: 'pip'
+    - run: pip install --upgrade pip
+    - run: pip install poetry
+    - run: poetry install --with dev
+    - run: poetry run pytest tests/
 
   run-tests-3_11:
     runs-on: ubuntu-latest
