+++ tests/helpers/util/exceptions.py
@@ -1,0 +1,2 @@
+async def awaitable_exception() -> Exception:
+    return Exception()
