+++ tests/routers/test_score_set.py
@@ -1,3 +1,5 @@
+# ruff: noqa: E402
+
 import re
 from copy import deepcopy
 from datetime import date
@@ -5,9 +7,12 @@ from unittest.mock import patch
 
 import jsonschema
 import pytest
-from arq import ArqRedis
 from humps import camelize
 from sqlalchemy import select
+
+arq = pytest.importorskip("arq")
+cdot = pytest.importorskip("cdot")
+fastapi = pytest.importorskip("fastapi")
 
 from mavedb.lib.validation.urn_re import MAVEDB_TMP_URN_RE, MAVEDB_SCORE_SET_URN_RE, MAVEDB_EXPERIMENT_URN_RE
 from mavedb.models.enums.processing_state import ProcessingState
@@ -16,6 +21,7 @@ from mavedb.models.score_set import ScoreSet as ScoreSetDbModel
 from mavedb.models.variant import Variant as VariantDbModel
 from mavedb.view_models.orcid import OrcidUser
 from mavedb.view_models.score_set import ScoreSet, ScoreSetCreate
+
 from tests.helpers.constants import (
     EXTRA_USER,
     EXTRA_LICENSE,
@@ -38,15 +44,13 @@ from tests.helpers.constants import (
     TEST_SAVED_SCORE_CALIBRATION,
 )
 from tests.helpers.dependency_overrider import DependencyOverrider
-from tests.helpers.util import (
-    add_contributor,
-    change_ownership,
-    change_to_inactive_license,
-    create_experiment,
-    create_seq_score_set,
-    create_seq_score_set_with_variants,
-    update_expected_response_for_created_resources,
-)
+from tests.helpers.util.contributor import add_contributor
+from tests.helpers.util.experiment import create_experiment
+from tests.helpers.util.license import change_to_inactive_license
+from tests.helpers.util.score_set import create_seq_score_set, publish_score_set
+from tests.helpers.util.common import update_expected_response_for_created_resources
+from tests.helpers.util.user import change_ownership
+from tests.helpers.util.variant import mock_worker_variant_insertion
 
 
 ########################################################################################################################
@@ -304,15 +308,12 @@ def test_can_update_score_set_supporting_data_after_publication(
     data_files,
 ):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        publication_response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
-        assert publication_response.status_code == 200
-        queue.assert_called_once()
-        published_score_set = publication_response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
     published_urn = published_score_set["urn"]
     response = client.get(f"/api/v1/score-sets/{published_urn}")
@@ -369,15 +370,12 @@ def test_cannot_update_score_set_target_data_after_publication(
     client, setup_router_db, attribute, expected_response_data, updated_data, session, data_provider, data_files
 ):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        publication_response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
-        assert publication_response.status_code == 200
-        queue.assert_called_once()
-        published_score_set = publication_response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
     published_urn = published_score_set["urn"]
     response = client.get(f"/api/v1/score-sets/{published_urn}")
@@ -535,7 +533,7 @@ def test_add_score_set_variants_scores_only_endpoint(client, setup_router_db, da
     scores_csv_path = data_files / "scores.csv"
     with (
         open(scores_csv_path, "rb") as scores_file,
-        patch.object(ArqRedis, "enqueue_job", return_value=None) as queue,
+        patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as queue,
     ):
         response = client.post(
             f"/api/v1/score-sets/{score_set['urn']}/variants/data",
@@ -561,7 +559,7 @@ def test_add_score_set_variants_scores_and_counts_endpoint(session, client, setu
     with (
         open(scores_csv_path, "rb") as scores_file,
         open(counts_csv_path, "rb") as counts_file,
-        patch.object(ArqRedis, "enqueue_job", return_value=None) as queue,
+        patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as queue,
     ):
         response = client.post(
             f"/api/v1/score-sets/{score_set['urn']}/variants/data",
@@ -588,7 +586,7 @@ def test_add_score_set_variants_scores_only_endpoint_utf8_encoded(client, setup_
     scores_csv_path = data_files / "scores_utf8_encoded.csv"
     with (
         open(scores_csv_path, "rb") as scores_file,
-        patch.object(ArqRedis, "enqueue_job", return_value=None) as queue,
+        patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as queue,
     ):
         response = client.post(
             f"/api/v1/score-sets/{score_set['urn']}/variants/data",
@@ -614,7 +612,7 @@ def test_add_score_set_variants_scores_and_counts_endpoint_utf8_encoded(session,
     with (
         open(scores_csv_path, "rb") as scores_file,
         open(counts_csv_path, "rb") as counts_file,
-        patch.object(ArqRedis, "enqueue_job", return_value=None) as queue,
+        patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as queue,
     ):
         response = client.post(
             f"/api/v1/score-sets/{score_set['urn']}/variants/data",
@@ -706,7 +704,7 @@ def test_contributor_can_add_scores_to_other_user_score_set(session, client, set
 
     with (
         open(scores_csv_path, "rb") as scores_file,
-        patch.object(ArqRedis, "enqueue_job", return_value=None) as queue,
+        patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as queue,
     ):
         response = client.post(
             f"/api/v1/score-sets/{score_set['urn']}/variants/data",
@@ -762,7 +760,7 @@ def test_contributor_can_add_scores_and_counts_to_other_user_score_set(session, 
     with (
         open(scores_csv_path, "rb") as scores_file,
         open(counts_csv_path, "rb") as counts_file,
-        patch.object(ArqRedis, "enqueue_job", return_value=None) as queue,
+        patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as queue,
     ):
         response = client.post(
             f"/api/v1/score-sets/{score_set['urn']}/variants/data",
@@ -813,7 +811,7 @@ def test_admin_can_add_scores_to_other_user_score_set(
     with (
         open(scores_csv_path, "rb") as scores_file,
         DependencyOverrider(admin_app_overrides),
-        patch.object(ArqRedis, "enqueue_job", return_value=None) as queue,
+        patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as queue,
     ):
         response = client.post(
             f"/api/v1/score-sets/{score_set['urn']}/variants/data",
@@ -839,7 +837,7 @@ def test_admin_can_add_scores_and_counts_to_other_user_score_set(session, client
     with (
         open(scores_csv_path, "rb") as scores_file,
         open(counts_csv_path, "rb") as counts_file,
-        patch.object(ArqRedis, "enqueue_job", return_value=None) as queue,
+        patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as queue,
     ):
         response = client.post(
             f"/api/v1/score-sets/{score_set['urn']}/variants/data",
@@ -867,26 +865,23 @@ def test_admin_can_add_scores_and_counts_to_other_user_score_set(session, client
 
 def test_publish_score_set(session, data_provider, client, setup_router_db, data_files):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        publication_response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
-        assert publication_response.status_code == 200
-        queue.assert_called_once()
-        response_data = publication_response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
-    assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(response_data["urn"]), re.Match)
-    assert isinstance(MAVEDB_EXPERIMENT_URN_RE.fullmatch(response_data["experiment"]["urn"]), re.Match)
+    assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(published_score_set["urn"]), re.Match)
+    assert isinstance(MAVEDB_EXPERIMENT_URN_RE.fullmatch(published_score_set["experiment"]["urn"]), re.Match)
 
     expected_response = update_expected_response_for_created_resources(
-        deepcopy(TEST_MINIMAL_SEQ_SCORESET_RESPONSE), response_data["experiment"], response_data
+        deepcopy(TEST_MINIMAL_SEQ_SCORESET_RESPONSE), published_score_set["experiment"], published_score_set
     )
     expected_response["experiment"].update({"publishedDate": date.today().isoformat()})
     expected_response.update(
         {
-            "urn": response_data["urn"],
+            "urn": published_score_set["urn"],
             "publishedDate": date.today().isoformat(),
             "numVariants": 3,
             "private": False,
@@ -894,10 +889,10 @@ def test_publish_score_set(session, data_provider, client, setup_router_db, data
             "processingState": ProcessingState.success.name,
         }
     )
-    assert sorted(expected_response.keys()) == sorted(response_data.keys())
+    assert sorted(expected_response.keys()) == sorted(published_score_set.keys())
 
     # refresh score set to post worker state
-    score_set = (client.get(f"/api/v1/score-sets/{response_data['urn']}")).json()
+    score_set = (client.get(f"/api/v1/score-sets/{published_score_set['urn']}")).json()
     for key in expected_response:
         assert (key, expected_response[key]) == (key, score_set[key])
 
@@ -909,27 +904,18 @@ def test_publish_score_set(session, data_provider, client, setup_router_db, data
 
 def test_publish_multiple_score_sets(session, data_provider, client, setup_router_db, data_files):
     experiment = create_experiment(client)
-    score_set_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv", update={"title": "Score Set 1"}
-    )
-    score_set_2 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv", update={"title": "Score Set 2"}
-    )
-    score_set_3 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv", update={"title": "Score Set 3"}
-    )
+    score_set_1 = create_seq_score_set(client, experiment["urn"])
+    score_set_1 = mock_worker_variant_insertion(client, session, data_provider, score_set_1, data_files / "scores.csv")
+    score_set_2 = create_seq_score_set(client, experiment["urn"])
+    score_set_2 = mock_worker_variant_insertion(client, session, data_provider, score_set_2, data_files / "scores.csv")
+    score_set_3 = create_seq_score_set(client, experiment["urn"])
+    score_set_3 = mock_worker_variant_insertion(client, session, data_provider, score_set_3, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        pub_score_set_1_response = client.post(f"/api/v1/score-sets/{score_set_1['urn']}/publish")
-        assert pub_score_set_1_response.status_code == 200
-        pub_score_set_2_response = client.post(f"/api/v1/score-sets/{score_set_2['urn']}/publish")
-        assert pub_score_set_2_response.status_code == 200
-        pub_score_set_3_response = client.post(f"/api/v1/score-sets/{score_set_3['urn']}/publish")
-        assert pub_score_set_3_response.status_code == 200
-        queue.assert_called()
-        pub_score_set_1_data = pub_score_set_1_response.json()
-        pub_score_set_2_data = pub_score_set_2_response.json()
-        pub_score_set_3_data = pub_score_set_3_response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        pub_score_set_1_data = publish_score_set(client, score_set_1["urn"])
+        pub_score_set_2_data = publish_score_set(client, score_set_2["urn"])
+        pub_score_set_3_data = publish_score_set(client, score_set_3["urn"])
+        worker_queue.assert_called()
 
     assert pub_score_set_1_data["urn"] == "urn:mavedb:00000001-a-1"
     assert pub_score_set_1_data["title"] == score_set_1["title"]
@@ -959,10 +945,10 @@ def test_cannot_publish_score_set_without_variants(client, setup_router_db):
     experiment = create_experiment(client)
     score_set = create_seq_score_set(client, experiment["urn"])
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
         response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
         assert response.status_code == 422
-        queue.assert_not_called()
+        worker_queue.assert_not_called()
         response_data = response.json()
 
     assert "cannot publish score set without variant scores" in response_data["detail"]
@@ -970,15 +956,15 @@ def test_cannot_publish_score_set_without_variants(client, setup_router_db):
 
 def test_cannot_publish_other_user_private_score_set(session, data_provider, client, setup_router_db, data_files):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
+
     change_ownership(session, score_set["urn"], ScoreSetDbModel)
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
         response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
         assert response.status_code == 404
-        queue.assert_not_called()
+        worker_queue.assert_not_called()
         response_data = response.json()
 
     assert f"score set with URN '{score_set['urn']}' not found" in response_data["detail"]
@@ -988,13 +974,12 @@ def test_anonymous_cannot_publish_user_private_score_set(
     session, data_provider, client, setup_router_db, data_files, anonymous_app_overrides
 ):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
     with (
         DependencyOverrider(anonymous_app_overrides),
-        patch.object(ArqRedis, "enqueue_job", return_value=None) as queue,
+        patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as queue,
     ):
         response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
         assert response.status_code == 401
@@ -1006,9 +991,8 @@ def test_anonymous_cannot_publish_user_private_score_set(
 
 def test_contributor_can_publish_other_users_score_set(session, data_provider, client, setup_router_db, data_files):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
     change_ownership(session, score_set["urn"], ScoreSetDbModel)
     add_contributor(
         session,
@@ -1019,22 +1003,20 @@ def test_contributor_can_publish_other_users_score_set(session, data_provider, c
         TEST_USER["last_name"],
     )
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
-        assert response.status_code == 200
-        queue.assert_called_once()
-        response_data = response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
-    assert response_data["urn"] == "urn:mavedb:00000001-a-1"
-    assert response_data["experiment"]["urn"] == "urn:mavedb:00000001-a"
+    assert published_score_set["urn"] == "urn:mavedb:00000001-a-1"
+    assert published_score_set["experiment"]["urn"] == "urn:mavedb:00000001-a"
 
     expected_response = update_expected_response_for_created_resources(
-        deepcopy(TEST_MINIMAL_SEQ_SCORESET_RESPONSE), response_data["experiment"], response_data
+        deepcopy(TEST_MINIMAL_SEQ_SCORESET_RESPONSE), published_score_set["experiment"], published_score_set
     )
     expected_response["experiment"].update({"publishedDate": date.today().isoformat()})
     expected_response.update(
         {
-            "urn": response_data["urn"],
+            "urn": published_score_set["urn"],
             "publishedDate": date.today().isoformat(),
             "numVariants": 3,
             "private": False,
@@ -1062,10 +1044,10 @@ def test_contributor_can_publish_other_users_score_set(session, data_provider, c
         "firstName": EXTRA_USER["first_name"],
         "lastName": EXTRA_USER["last_name"],
     }
-    assert sorted(expected_response.keys()) == sorted(response_data.keys())
+    assert sorted(expected_response.keys()) == sorted(published_score_set.keys())
 
     # refresh score set to post worker state
-    score_set = (client.get(f"/api/v1/score-sets/{response_data['urn']}")).json()
+    score_set = (client.get(f"/api/v1/score-sets/{published_score_set['urn']}")).json()
     for key in expected_response:
         assert (key, expected_response[key]) == (key, score_set[key])
 
@@ -1079,11 +1061,13 @@ def test_admin_cannot_publish_other_user_private_score_set(
     session, data_provider, client, admin_app_overrides, setup_router_db, data_files
 ):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with DependencyOverrider(admin_app_overrides), patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
+    with (
+        DependencyOverrider(admin_app_overrides),
+        patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as queue,
+    ):
         response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
         assert response.status_code == 404
         queue.assert_not_called()
@@ -1099,57 +1083,49 @@ def test_admin_cannot_publish_other_user_private_score_set(
 
 def test_create_single_score_set_meta_analysis(session, data_provider, client, setup_router_db, data_files):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
-        assert response.status_code == 200
-        queue.assert_called_once()
-        score_set = response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
-    meta_score_set = create_seq_score_set_with_variants(
+    meta_score_set = create_seq_score_set(
         client,
-        session,
-        data_provider,
         None,
-        data_files / "scores.csv",
-        update={"title": "Test Meta Analysis", "metaAnalyzesScoreSetUrns": [score_set["urn"]]},
+        update={"title": "Test Meta Analysis", "metaAnalyzesScoreSetUrns": [published_score_set["urn"]]},
+    )
+    meta_score_set = mock_worker_variant_insertion(
+        client, session, data_provider, meta_score_set, data_files / "scores.csv"
     )
 
-    score_set_refresh = (client.get(f"/api/v1/score-sets/{score_set['urn']}")).json()
-    assert meta_score_set["metaAnalyzesScoreSetUrns"] == [score_set["urn"]]
-    assert score_set_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set["urn"]]
+    published_score_set_refresh = (client.get(f"/api/v1/score-sets/{published_score_set['urn']}")).json()
+    assert meta_score_set["metaAnalyzesScoreSetUrns"] == [published_score_set_refresh["urn"]]
+    assert published_score_set_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set["urn"]]
     assert isinstance(MAVEDB_TMP_URN_RE.fullmatch(meta_score_set["urn"]), re.Match)
 
 
 def test_publish_single_score_set_meta_analysis(session, data_provider, client, setup_router_db, data_files):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
-        assert response.status_code == 200
-        queue.assert_called_once()
-        score_set = response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
-    meta_score_set = create_seq_score_set_with_variants(
+    meta_score_set = create_seq_score_set(
         client,
-        session,
-        data_provider,
         None,
-        data_files / "scores.csv",
         update={"title": "Test Meta Analysis", "metaAnalyzesScoreSetUrns": [score_set["urn"]]},
     )
+    meta_score_set = mock_worker_variant_insertion(
+        client, session, data_provider, meta_score_set, data_files / "scores.csv"
+    )
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        meta_response = client.post(f"/api/v1/score-sets/{meta_score_set['urn']}/publish")
-        assert meta_response.status_code == 200
-        queue.assert_called_once()
-        meta_score_set = meta_response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        meta_score_set = publish_score_set(client, meta_score_set["urn"])
+        worker_queue.assert_called_once()
 
     assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(meta_score_set["urn"]), re.Match)
     assert meta_score_set["urn"] == "urn:mavedb:00000001-0-1"
@@ -1159,42 +1135,38 @@ def test_multiple_score_set_meta_analysis_single_experiment(
     session, data_provider, client, setup_router_db, data_files
 ):
     experiment = create_experiment(client)
-    score_set_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv", update={"title": "Score Set 1"}
-    )
-    score_set_2 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv", update={"title": "Score Set 2"}
-    )
+    score_set_1 = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set_1 = mock_worker_variant_insertion(client, session, data_provider, score_set_1, data_files / "scores.csv")
+    score_set_2 = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 2"})
+    score_set_2 = mock_worker_variant_insertion(client, session, data_provider, score_set_2, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        response_1 = client.post(f"/api/v1/score-sets/{score_set_1['urn']}/publish")
-        assert response_1.status_code == 200
-        response_2 = client.post(f"/api/v1/score-sets/{score_set_2['urn']}/publish")
-        assert response_2.status_code == 200
-        queue.assert_called()
-        score_set_1 = response_1.json()
-        score_set_2 = response_2.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set_1 = publish_score_set(client, score_set_1["urn"])
+        published_score_set_2 = publish_score_set(client, score_set_2["urn"])
+        worker_queue.assert_called()
 
-    meta_score_set = create_seq_score_set_with_variants(
+    meta_score_set = create_seq_score_set(
         client,
-        session,
-        data_provider,
         None,
-        data_files / "scores.csv",
-        update={"title": "Test Meta Analysis", "metaAnalyzesScoreSetUrns": [score_set_1["urn"], score_set_2["urn"]]},
+        update={
+            "title": "Test Meta Analysis",
+            "metaAnalyzesScoreSetUrns": [published_score_set_1["urn"], published_score_set_2["urn"]],
+        },
     )
-    score_set_1_refresh = (client.get(f"/api/v1/score-sets/{score_set_1['urn']}")).json()
-    assert meta_score_set["metaAnalyzesScoreSetUrns"] == sorted([score_set_1["urn"], score_set_2["urn"]])
-    assert score_set_1_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set["urn"]]
+    meta_score_set = mock_worker_variant_insertion(
+        client, session, data_provider, meta_score_set, data_files / "scores.csv"
+    )
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        meta_response = client.post(f"/api/v1/score-sets/{meta_score_set['urn']}/publish")
-        assert meta_response.status_code == 200
-        queue.assert_called_once()
-        meta_score_set = meta_response.json()
+    published_score_set_1_refresh = (client.get(f"/api/v1/score-sets/{published_score_set_1['urn']}")).json()
+    assert meta_score_set["metaAnalyzesScoreSetUrns"] == sorted([published_score_set_1["urn"], published_score_set_2["urn"]])
+    assert published_score_set_1_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set["urn"]]
 
-    assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(meta_score_set["urn"]), re.Match)
-    assert meta_score_set["urn"] == "urn:mavedb:00000001-0-1"
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_meta_score_set = publish_score_set(client, meta_score_set["urn"])
+        worker_queue.assert_called_once()
+
+    assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(published_meta_score_set["urn"]), re.Match)
+    assert published_meta_score_set["urn"] == "urn:mavedb:00000001-0-1"
 
 
 def test_multiple_score_set_meta_analysis_multiple_experiment_sets(
@@ -1202,42 +1174,39 @@ def test_multiple_score_set_meta_analysis_multiple_experiment_sets(
 ):
     experiment_1 = create_experiment(client, {"title": "Experiment 1"})
     experiment_2 = create_experiment(client, {"title": "Experiment 2"})
-    score_set_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment_1["urn"], data_files / "scores.csv", update={"title": "Score Set 1"}
-    )
-    score_set_2 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment_2["urn"], data_files / "scores.csv", update={"title": "Score Set 2"}
-    )
+    score_set_1 = create_seq_score_set(client, experiment_1["urn"], update={"title": "Score Set 1"})
+    score_set_1 = mock_worker_variant_insertion(client, session, data_provider, score_set_1, data_files / "scores.csv")
+    score_set_2 = create_seq_score_set(client, experiment_2["urn"], update={"title": "Score Set 2"})
+    score_set_2 = mock_worker_variant_insertion(client, session, data_provider, score_set_2, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        response_1 = client.post(f"/api/v1/score-sets/{score_set_1['urn']}/publish")
-        assert response_1.status_code == 200
-        response_2 = client.post(f"/api/v1/score-sets/{score_set_2['urn']}/publish")
-        assert response_2.status_code == 200
-        queue.assert_called()
-        score_set_1 = response_1.json()
-        score_set_2 = response_2.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set_1 = publish_score_set(client, score_set_1["urn"])
+        published_score_set_2 = publish_score_set(client, score_set_2["urn"])
+        worker_queue.assert_called()
 
-    meta_score_set = create_seq_score_set_with_variants(
+    meta_score_set = create_seq_score_set(
         client,
-        session,
-        data_provider,
         None,
-        data_files / "scores.csv",
-        update={"title": "Test Meta Analysis", "metaAnalyzesScoreSetUrns": [score_set_1["urn"], score_set_2["urn"]]},
+        update={
+            "title": "Test Meta Analysis",
+            "metaAnalyzesScoreSetUrns": [published_score_set_1["urn"], published_score_set_2["urn"]],
+        },
     )
-    score_set_1_refresh = (client.get(f"/api/v1/score-sets/{score_set_1['urn']}")).json()
-    assert meta_score_set["metaAnalyzesScoreSetUrns"] == sorted([score_set_1["urn"], score_set_2["urn"]])
-    assert score_set_1_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set["urn"]]
+    meta_score_set = mock_worker_variant_insertion(
+        client, session, data_provider, meta_score_set, data_files / "scores.csv"
+    )
+    published_score_set_1_refresh = (client.get(f"/api/v1/score-sets/{published_score_set_1['urn']}")).json()
+    assert meta_score_set["metaAnalyzesScoreSetUrns"] == sorted(
+        [published_score_set_1["urn"], published_score_set_2["urn"]]
+    )
+    assert published_score_set_1_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set["urn"]]
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        meta_response = client.post(f"/api/v1/score-sets/{meta_score_set['urn']}/publish")
-        assert meta_response.status_code == 200
-        queue.assert_called_once()
-        meta_score_set = meta_response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_meta_score_set = publish_score_set(client, meta_score_set["urn"])
+        worker_queue.assert_called_once()
 
-    assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(meta_score_set["urn"]), re.Match)
-    assert meta_score_set["urn"] == "urn:mavedb:00000003-0-1"
+    assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(published_meta_score_set["urn"]), re.Match)
+    assert published_meta_score_set["urn"] == "urn:mavedb:00000003-0-1"
 
 
 def test_multiple_score_set_meta_analysis_multiple_experiments(
@@ -1247,42 +1216,39 @@ def test_multiple_score_set_meta_analysis_multiple_experiments(
     experiment_2 = create_experiment(
         client, {"title": "Experiment 2", "experimentSetUrn": experiment_1["experimentSetUrn"]}
     )
-    score_set_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment_1["urn"], data_files / "scores.csv", update={"title": "Score Set 1"}
-    )
-    score_set_2 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment_2["urn"], data_files / "scores.csv", update={"title": "Score Set 2"}
-    )
+    score_set_1 = create_seq_score_set(client, experiment_1["urn"], update={"title": "Score Set 1"})
+    score_set_1 = mock_worker_variant_insertion(client, session, data_provider, score_set_1, data_files / "scores.csv")
+    score_set_2 = create_seq_score_set(client, experiment_2["urn"], update={"title": "Score Set 2"})
+    score_set_2 = mock_worker_variant_insertion(client, session, data_provider, score_set_2, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        response_1 = client.post(f"/api/v1/score-sets/{score_set_1['urn']}/publish")
-        assert response_1.status_code == 200
-        response_2 = client.post(f"/api/v1/score-sets/{score_set_2['urn']}/publish")
-        assert response_2.status_code == 200
-        queue.assert_called()
-        score_set_1 = response_1.json()
-        score_set_2 = response_2.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set_1 = publish_score_set(client, score_set_1["urn"])
+        published_score_set_2 = publish_score_set(client, score_set_2["urn"])
+        worker_queue.assert_called()
 
-    meta_score_set = create_seq_score_set_with_variants(
+    meta_score_set = create_seq_score_set(
         client,
-        session,
-        data_provider,
         None,
-        data_files / "scores.csv",
-        update={"title": "Test Meta Analysis", "metaAnalyzesScoreSetUrns": [score_set_1["urn"], score_set_2["urn"]]},
+        update={
+            "title": "Test Meta Analysis",
+            "metaAnalyzesScoreSetUrns": [published_score_set_1["urn"], published_score_set_2["urn"]],
+        },
     )
-    score_set_1_refresh = (client.get(f"/api/v1/score-sets/{score_set_1['urn']}")).json()
-    assert meta_score_set["metaAnalyzesScoreSetUrns"] == sorted([score_set_1["urn"], score_set_2["urn"]])
-    assert score_set_1_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set["urn"]]
+    meta_score_set = mock_worker_variant_insertion(
+        client, session, data_provider, meta_score_set, data_files / "scores.csv"
+    )
+    published_score_set_1_refresh = (client.get(f"/api/v1/score-sets/{published_score_set_1['urn']}")).json()
+    assert meta_score_set["metaAnalyzesScoreSetUrns"] == sorted(
+        [published_score_set_1["urn"], published_score_set_2["urn"]]
+    )
+    assert published_score_set_1_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set["urn"]]
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        meta_response = client.post(f"/api/v1/score-sets/{meta_score_set['urn']}/publish")
-        assert meta_response.status_code == 200
-        queue.assert_called_once()
-        meta_score_set = meta_response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_meta_score_set = publish_score_set(client, meta_score_set["urn"])
+        worker_queue.assert_called_once()
 
-    assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(meta_score_set["urn"]), re.Match)
-    assert meta_score_set["urn"] == "urn:mavedb:00000001-0-1"
+    assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(published_meta_score_set["urn"]), re.Match)
+    assert published_meta_score_set["urn"] == "urn:mavedb:00000001-0-1"
 
 
 def test_multiple_score_set_meta_analysis_multiple_experiment_sets_different_score_sets(
@@ -1290,133 +1256,117 @@ def test_multiple_score_set_meta_analysis_multiple_experiment_sets_different_sco
 ):
     experiment_1 = create_experiment(client, {"title": "Experiment 1"})
     experiment_2 = create_experiment(client, {"title": "Experiment 2"})
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_1["urn"],
-        data_files / "scores.csv",
-        update={"title": "Exp 1 Score Set 1"},
+
+    score_set_1_1 = create_seq_score_set(client, experiment_1["urn"], update={"title": "Score Set 1 exp 1"})
+    score_set_1_1 = mock_worker_variant_insertion(
+        client, session, data_provider, score_set_1_1, data_files / "scores.csv"
     )
-    score_set_1_2 = create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_1["urn"],
-        data_files / "scores.csv",
-        update={"title": "Exp 1 Score Set 2"},
+    score_set_2_1 = create_seq_score_set(client, experiment_1["urn"], update={"title": "Score Set 2 exp 1"})
+    score_set_2_1 = mock_worker_variant_insertion(
+        client, session, data_provider, score_set_2_1, data_files / "scores.csv"
     )
-    score_set_2_1 = create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_2["urn"],
-        data_files / "scores.csv",
-        update={"title": "Exp 2 Score Set 1"},
+    score_set_1_2 = create_seq_score_set(client, experiment_2["urn"], update={"title": "Score Set 1 exp 2 "})
+    score_set_1_2 = mock_worker_variant_insertion(
+        client, session, data_provider, score_set_1_2, data_files / "scores.csv"
     )
-    score_set_2_2 = create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_2["urn"],
-        data_files / "scores.csv",
-        update={"title": "Exp 2 Score Set 2"},
+    score_set_2_2 = create_seq_score_set(client, experiment_2["urn"], update={"title": "Score Set 2 exp 2"})
+    score_set_2_2 = mock_worker_variant_insertion(
+        client, session, data_provider, score_set_2_2, data_files / "scores.csv"
     )
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        response_1_1 = client.post(f"/api/v1/score-sets/{score_set_1_1['urn']}/publish")
-        assert response_1_1.status_code == 200
-        response_1_2 = client.post(f"/api/v1/score-sets/{score_set_1_2['urn']}/publish")
-        assert response_1_2.status_code == 200
-        response_2_1 = client.post(f"/api/v1/score-sets/{score_set_2_1['urn']}/publish")
-        assert response_2_1.status_code == 200
-        response_2_2 = client.post(f"/api/v1/score-sets/{score_set_2_2['urn']}/publish")
-        assert response_2_2.status_code == 200
-        queue.assert_called()
-        score_set_1_1 = response_1_1.json()
-        score_set_1_2 = response_1_2.json()
-        score_set_2_1 = response_2_1.json()
-        score_set_2_2 = response_2_2.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set_1_1 = publish_score_set(client, score_set_1_1["urn"])
+        published_score_set_1_2 = publish_score_set(client, score_set_1_2["urn"])
+        published_score_set_2_1 = publish_score_set(client, score_set_2_1["urn"])
+        published_score_set_2_2 = publish_score_set(client, score_set_2_2["urn"])
+        worker_queue.assert_called()
 
-    meta_score_set_1 = create_seq_score_set_with_variants(
+    meta_score_set_1 = create_seq_score_set(
         client,
-        session,
-        data_provider,
         None,
-        data_files / "scores.csv",
         update={
-            "title": "Test Meta Analysis 1-1 2-1",
-            "metaAnalyzesScoreSetUrns": [score_set_1_1["urn"], score_set_2_1["urn"]],
+            "title": "Test Meta Analysis",
+            "metaAnalyzesScoreSetUrns": [published_score_set_1_1["urn"], published_score_set_1_2["urn"]],
         },
     )
-    score_set_1_1_refresh = (client.get(f"/api/v1/score-sets/{score_set_1_1['urn']}")).json()
-    assert meta_score_set_1["metaAnalyzesScoreSetUrns"] == sorted([score_set_1_1["urn"], score_set_2_1["urn"]])
-    assert score_set_1_1_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set_1["urn"]]
-    meta_score_set_2 = create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        None,
-        data_files / "scores.csv",
-        update={
-            "title": "Test Meta Analysis 1-2 2-2",
-            "metaAnalyzesScoreSetUrns": [score_set_1_2["urn"], score_set_2_2["urn"]],
-        },
+    meta_score_set_1 = mock_worker_variant_insertion(
+        client, session, data_provider, meta_score_set_1, data_files / "scores.csv"
     )
 
-    meta_score_set_3 = create_seq_score_set_with_variants(
+    published_score_set_1_1_refresh = (client.get(f"/api/v1/score-sets/{published_score_set_1_1['urn']}")).json()
+    assert meta_score_set_1["metaAnalyzesScoreSetUrns"] == sorted(
+        [published_score_set_1_1["urn"], published_score_set_1_2["urn"]]
+    )
+    assert published_score_set_1_1_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set_1["urn"]]
+
+    meta_score_set_2 = create_seq_score_set(
         client,
-        session,
-        data_provider,
         None,
-        data_files / "scores.csv",
         update={
-            "title": "Test Meta Analysis 1-1 2-2",
-            "metaAnalyzesScoreSetUrns": [score_set_1_1["urn"], score_set_2_2["urn"]],
+            "title": "Test Meta Analysis",
+            "metaAnalyzesScoreSetUrns": [published_score_set_2_1["urn"], published_score_set_2_2["urn"]],
         },
     )
+    meta_score_set_2 = mock_worker_variant_insertion(
+        client, session, data_provider, meta_score_set_2, data_files / "scores.csv"
+    )
+    published_score_set_2_1_refresh = (client.get(f"/api/v1/score-sets/{published_score_set_2_1['urn']}")).json()
+    assert meta_score_set_2["metaAnalyzesScoreSetUrns"] == sorted(
+        [published_score_set_2_1["urn"], published_score_set_2_2["urn"]]
+    )
+    assert published_score_set_2_1_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set_2["urn"]]
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        meta_score_set_1 = (client.post(f"/api/v1/score-sets/{meta_score_set_1['urn']}/publish")).json()
-        assert meta_score_set_1["urn"] == "urn:mavedb:00000003-0-1"
-        meta_score_set_2 = (client.post(f"/api/v1/score-sets/{meta_score_set_2['urn']}/publish")).json()
-        assert meta_score_set_2["urn"] == "urn:mavedb:00000003-0-2"
-        meta_score_set_3 = (client.post(f"/api/v1/score-sets/{meta_score_set_3['urn']}/publish")).json()
-        assert meta_score_set_3["urn"] == "urn:mavedb:00000003-0-3"
-        queue.assert_called()
+    meta_score_set_3 = create_seq_score_set(
+        client,
+        None,
+        update={
+            "title": "Test Meta Analysis",
+            "metaAnalyzesScoreSetUrns": [published_score_set_1_1["urn"], published_score_set_2_2["urn"]],
+        },
+    )
+    meta_score_set_3 = mock_worker_variant_insertion(
+        client, session, data_provider, meta_score_set_3, data_files / "scores.csv"
+    )
 
-    assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(meta_score_set_1["urn"]), re.Match)
-    assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(meta_score_set_2["urn"]), re.Match)
-    assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(meta_score_set_3["urn"]), re.Match)
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_meta_score_set_1 = publish_score_set(client, meta_score_set_1["urn"])
+        published_meta_score_set_2 = publish_score_set(client, meta_score_set_2["urn"])
+        published_meta_score_set_3 = publish_score_set(client, meta_score_set_3["urn"])
+        worker_queue.assert_called()
+
+    assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(published_meta_score_set_1["urn"]), re.Match)
+    assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(published_meta_score_set_2["urn"]), re.Match)
+    assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(published_meta_score_set_3["urn"]), re.Match)
+    assert published_meta_score_set_1["urn"] == "urn:mavedb:00000003-0-1"
+    assert published_meta_score_set_2["urn"] == "urn:mavedb:00000003-0-2"
+    assert published_meta_score_set_3["urn"] == "urn:mavedb:00000003-0-3"
 
 
 def test_cannot_add_score_set_to_meta_analysis_experiment(session, data_provider, client, setup_router_db, data_files):
     experiment = create_experiment(client)
-    score_set_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set_1 = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set_1 = mock_worker_variant_insertion(client, session, data_provider, score_set_1, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        response = client.post(f"/api/v1/score-sets/{score_set_1['urn']}/publish")
-        assert response.status_code == 200
-        queue.assert_called_once()
-        score_set_1 = response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set_1 = publish_score_set(client, score_set_1["urn"])
+        worker_queue.assert_called()
 
-    meta_score_set_1 = create_seq_score_set_with_variants(
+    meta_score_set_1 = create_seq_score_set(
         client,
-        session,
-        data_provider,
         None,
-        data_files / "scores.csv",
-        update={"title": "Test Meta Analysis", "metaAnalyzesScoreSetUrns": [score_set_1["urn"]]},
+        update={"title": "Test Meta Analysis", "metaAnalyzesScoreSetUrns": [published_score_set_1["urn"]]},
+    )
+    meta_score_set_1 = mock_worker_variant_insertion(
+        client, session, data_provider, meta_score_set_1, data_files / "scores.csv"
     )
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        meta_score_set_1 = (client.post(f"/api/v1/score-sets/{meta_score_set_1['urn']}/publish")).json()
-        assert meta_score_set_1["urn"] == "urn:mavedb:00000001-0-1"
-        queue.assert_called()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        meta_score_set_1 = publish_score_set(client, meta_score_set_1["urn"])
+        worker_queue.assert_called()
 
     assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(meta_score_set_1["urn"]), re.Match)
+    assert meta_score_set_1["urn"] == "urn:mavedb:00000001-0-1"
+
     score_set_2 = deepcopy(TEST_MINIMAL_SEQ_SCORESET)
     score_set_2["experimentUrn"] = meta_score_set_1["experiment"]["urn"]
     jsonschema.validate(instance=score_set_2, schema=ScoreSetCreate.schema())
@@ -1431,29 +1381,27 @@ def test_create_single_score_set_meta_analysis_to_others_score_set(
     session, data_provider, client, setup_router_db, data_files
 ):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
-        assert response.status_code == 200
-        queue.assert_called_once()
-        score_set = response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called()
 
-    change_ownership(session, score_set["urn"], ScoreSetDbModel)
-    meta_score_set = create_seq_score_set_with_variants(
+    change_ownership(session, published_score_set["urn"], ScoreSetDbModel)
+
+    meta_score_set = create_seq_score_set(
         client,
-        session,
-        data_provider,
         None,
-        data_files / "scores.csv",
-        update={"title": "Test Meta Analysis", "metaAnalyzesScoreSetUrns": [score_set["urn"]]},
+        update={"title": "Test Meta Analysis", "metaAnalyzesScoreSetUrns": [published_score_set["urn"]]},
+    )
+    meta_score_set = mock_worker_variant_insertion(
+        client, session, data_provider, meta_score_set, data_files / "scores.csv"
     )
 
-    score_set_refresh = (client.get(f"/api/v1/score-sets/{score_set['urn']}")).json()
-    assert meta_score_set["metaAnalyzesScoreSetUrns"] == [score_set["urn"]]
-    assert score_set_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set["urn"]]
+    published_score_set_refresh = (client.get(f"/api/v1/score-sets/{published_score_set['urn']}")).json()
+    assert meta_score_set["metaAnalyzesScoreSetUrns"] == [published_score_set["urn"]]
+    assert published_score_set_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set["urn"]]
     assert isinstance(MAVEDB_TMP_URN_RE.fullmatch(meta_score_set["urn"]), re.Match)
 
 
@@ -1461,40 +1409,38 @@ def test_multiple_score_set_meta_analysis_single_experiment_with_different_creat
     session, data_provider, client, setup_router_db, data_files
 ):
     experiment = create_experiment(client)
-    score_set_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv", update={"title": "Score Set 1"}
-    )
-    score_set_2 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv", update={"title": "Score Set 2"}
-    )
+    score_set_1 = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set_1 = mock_worker_variant_insertion(client, session, data_provider, score_set_1, data_files / "scores.csv")
+    score_set_2 = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 2"})
+    score_set_2 = mock_worker_variant_insertion(client, session, data_provider, score_set_2, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        response_1 = client.post(f"/api/v1/score-sets/{score_set_1['urn']}/publish")
-        assert response_1.status_code == 200
-        response_2 = client.post(f"/api/v1/score-sets/{score_set_2['urn']}/publish")
-        assert response_2.status_code == 200
-        queue.assert_called()
-        score_set_1 = response_1.json()
-        score_set_2 = response_2.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set_1 = publish_score_set(client, score_set_1["urn"])
+        published_score_set_2 = publish_score_set(client, score_set_2["urn"])
+        worker_queue.assert_called()
 
-    change_ownership(session, score_set_2["urn"], ScoreSetDbModel)
-    meta_score_set = create_seq_score_set_with_variants(
+    change_ownership(session, published_score_set_2["urn"], ScoreSetDbModel)
+    meta_score_set = create_seq_score_set(
         client,
-        session,
-        data_provider,
         None,
-        data_files / "scores.csv",
-        update={"title": "Test Meta Analysis", "metaAnalyzesScoreSetUrns": [score_set_1["urn"], score_set_2["urn"]]},
+        update={
+            "title": "Test Meta Analysis",
+            "metaAnalyzesScoreSetUrns": [published_score_set_1["urn"], published_score_set_2["urn"]],
+        },
     )
-    score_set_1_refresh = (client.get(f"/api/v1/score-sets/{score_set_1['urn']}")).json()
-    assert meta_score_set["metaAnalyzesScoreSetUrns"] == sorted([score_set_1["urn"], score_set_2["urn"]])
-    assert score_set_1_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set["urn"]]
+    meta_score_set = mock_worker_variant_insertion(
+        client, session, data_provider, meta_score_set, data_files / "scores.csv"
+    )
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        meta_response = client.post(f"/api/v1/score-sets/{meta_score_set['urn']}/publish")
-        assert meta_response.status_code == 200
-        queue.assert_called_once()
-        meta_score_set = meta_response.json()
+    published_score_set_1_refresh = (client.get(f"/api/v1/score-sets/{published_score_set_1['urn']}")).json()
+    assert meta_score_set["metaAnalyzesScoreSetUrns"] == sorted(
+        [published_score_set_1["urn"], published_score_set_2["urn"]]
+    )
+    assert published_score_set_1_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set["urn"]]
+
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        meta_score_set = publish_score_set(client, meta_score_set["urn"])
+        worker_queue.assert_called()
 
     assert meta_score_set["urn"] == "urn:mavedb:00000001-0-1"
     assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(meta_score_set["urn"]), re.Match)
@@ -1505,39 +1451,41 @@ def test_multiple_score_set_meta_analysis_multiple_experiment_sets_with_differen
 ):
     experiment_1 = create_experiment(client, {"title": "Experiment 1"})
     experiment_2 = create_experiment(client, {"title": "Experiment 2"})
-    score_set_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment_1["urn"], data_files / "scores.csv", update={"title": "Score Set 1"}
-    )
-    score_set_2 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment_2["urn"], data_files / "scores.csv", update={"title": "Score Set 2"}
-    )
+    score_set_1 = create_seq_score_set(client, experiment_1["urn"], update={"title": "Score Set 1"})
+    score_set_1 = mock_worker_variant_insertion(client, session, data_provider, score_set_1, data_files / "scores.csv")
+    score_set_2 = create_seq_score_set(client, experiment_2["urn"], update={"title": "Score Set 2"})
+    score_set_2 = mock_worker_variant_insertion(client, session, data_provider, score_set_2, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        score_set_1 = (client.post(f"/api/v1/score-sets/{score_set_1['urn']}/publish")).json()
-        score_set_2 = (client.post(f"/api/v1/score-sets/{score_set_2['urn']}/publish")).json()
-        queue.assert_called()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set_1 = publish_score_set(client, score_set_1["urn"])
+        published_score_set_2 = publish_score_set(client, score_set_2["urn"])
+        worker_queue.assert_called()
 
-    change_ownership(session, score_set_2["urn"], ScoreSetDbModel)
-    meta_score_set = create_seq_score_set_with_variants(
+    change_ownership(session, published_score_set_2["urn"], ScoreSetDbModel)
+    meta_score_set = create_seq_score_set(
         client,
-        session,
-        data_provider,
         None,
-        data_files / "scores.csv",
-        update={"title": "Test Meta Analysis", "metaAnalyzesScoreSetUrns": [score_set_1["urn"], score_set_2["urn"]]},
+        update={
+            "title": "Test Meta Analysis",
+            "metaAnalyzesScoreSetUrns": [published_score_set_1["urn"], published_score_set_2["urn"]],
+        },
     )
-    score_set_1_refresh = (client.get(f"/api/v1/score-sets/{score_set_1['urn']}")).json()
-    assert meta_score_set["metaAnalyzesScoreSetUrns"] == sorted([score_set_1["urn"], score_set_2["urn"]])
-    assert score_set_1_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set["urn"]]
+    meta_score_set = mock_worker_variant_insertion(
+        client, session, data_provider, meta_score_set, data_files / "scores.csv"
+    )
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        meta_response = client.post(f"/api/v1/score-sets/{meta_score_set['urn']}/publish")
-        assert meta_response.status_code == 200
-        queue.assert_called_once()
-        meta_score_set = meta_response.json()
+    published_score_set_1_refresh = (client.get(f"/api/v1/score-sets/{published_score_set_1['urn']}")).json()
+    assert meta_score_set["metaAnalyzesScoreSetUrns"] == sorted(
+        [published_score_set_1["urn"], published_score_set_2["urn"]]
+    )
+    assert published_score_set_1_refresh["metaAnalyzedByScoreSetUrns"] == [meta_score_set["urn"]]
 
-    assert meta_score_set["urn"] == "urn:mavedb:00000003-0-1"
-    assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(meta_score_set["urn"]), re.Match)
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_meta_score_set = publish_score_set(client, meta_score_set["urn"])
+        worker_queue.assert_called()
+
+    assert published_meta_score_set["urn"] == "urn:mavedb:00000003-0-1"
+    assert isinstance(MAVEDB_SCORE_SET_URN_RE.fullmatch(published_meta_score_set["urn"]), re.Match)
 
 
 ########################################################################################################################
@@ -1546,15 +1494,9 @@ def test_multiple_score_set_meta_analysis_multiple_experiment_sets_with_differen
 
 
 def test_search_private_score_sets_no_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client, {"title": "Experiment 1"})
-    create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_1["urn"],
-        data_files / "scores.csv",
-        update={"title": "Test Score Set"},
-    )
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
     search_payload = {"text": "fnord"}
     response = client.post("/api/v1/me/score-sets/search", json=search_payload)
@@ -1563,61 +1505,49 @@ def test_search_private_score_sets_no_match(session, data_provider, client, setu
 
 
 def test_search_private_score_sets_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client, {"title": "Experiment 1"})
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_1["urn"],
-        data_files / "scores.csv",
-        update={"title": "Test Fnord Score Set"},
-    )
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"], update={"title": "Test Fnord Score Set"})
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
     search_payload = {"text": "fnord"}
     response = client.post("/api/v1/me/score-sets/search", json=search_payload)
     assert response.status_code == 200
     assert len(response.json()) == 1
-    assert response.json()[0]["title"] == score_set_1_1["title"]
+    assert response.json()[0]["title"] == score_set["title"]
 
 
 def test_search_private_score_sets_urn_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client)
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment_1["urn"], data_files / "scores.csv"
-    )
+    experiment = create_experiment(client)
+    score_set = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    search_payload = {"urn": score_set_1_1["urn"]}
+    search_payload = {"urn": score_set["urn"]}
     response = client.post("/api/v1/me/score-sets/search", json=search_payload)
     assert response.status_code == 200
     assert len(response.json()) == 1
-    assert response.json()[0]["urn"] == score_set_1_1["urn"]
+    assert response.json()[0]["urn"] == score_set["urn"]
 
 
 # There is space in the end of test urn. The search result returned nothing before.
 def test_search_private_score_sets_urn_with_space_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client)
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment_1["urn"], data_files / "scores.csv"
-    )
-    urn_with_space = score_set_1_1["urn"] + "   "
+    experiment = create_experiment(client)
+    score_set = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
+
+    urn_with_space = score_set["urn"] + "   "
     search_payload = {"urn": urn_with_space}
     response = client.post("/api/v1/me/score-sets/search", json=search_payload)
     assert response.status_code == 200
     assert len(response.json()) == 1
-    assert response.json()[0]["urn"] == score_set_1_1["urn"]
+    assert response.json()[0]["urn"] == score_set["urn"]
 
 
 def test_search_others_private_score_sets_no_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client, {"title": "Experiment 1"})
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_1["urn"],
-        data_files / "scores.csv",
-        update={"title": "Test Score Set"},
-    )
-    change_ownership(session, score_set_1_1["urn"], ScoreSetDbModel)
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
+    change_ownership(session, score_set["urn"], ScoreSetDbModel)
+
     search_payload = {"text": "fnord"}
     response = client.post("/api/v1/me/score-sets/search", json=search_payload)
     assert response.status_code == 200
@@ -1625,16 +1555,11 @@ def test_search_others_private_score_sets_no_match(session, data_provider, clien
 
 
 def test_search_others_private_score_sets_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client, {"title": "Experiment 1"})
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_1["urn"],
-        data_files / "scores.csv",
-        update={"title": "Test Fnord Score Set"},
-    )
-    change_ownership(session, score_set_1_1["urn"], ScoreSetDbModel)
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
+
+    change_ownership(session, score_set["urn"], ScoreSetDbModel)
     search_payload = {"text": "fnord"}
     response = client.post("/api/v1/me/score-sets/search", json=search_payload)
     assert response.status_code == 200
@@ -1642,12 +1567,12 @@ def test_search_others_private_score_sets_match(session, data_provider, client, 
 
 
 def test_search_others_private_score_sets_urn_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client)
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment_1["urn"], data_files / "scores.csv"
-    )
-    change_ownership(session, score_set_1_1["urn"], ScoreSetDbModel)
-    search_payload = {"urn": score_set_1_1["urn"]}
+    experiment = create_experiment(client)
+    score_set = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
+    change_ownership(session, score_set["urn"], ScoreSetDbModel)
+
+    search_payload = {"urn": score_set["urn"]}
     response = client.post("/api/v1/me/score-sets/search", json=search_payload)
     assert response.status_code == 200
     assert len(response.json()) == 0
@@ -1657,12 +1582,12 @@ def test_search_others_private_score_sets_urn_match(session, data_provider, clie
 def test_search_others_private_score_sets_urn_with_space_match(
     session, data_provider, client, setup_router_db, data_files
 ):
-    experiment_1 = create_experiment(client)
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment_1["urn"], data_files / "scores.csv"
-    )
-    change_ownership(session, score_set_1_1["urn"], ScoreSetDbModel)
-    urn_with_space = score_set_1_1["urn"] + "   "
+    experiment = create_experiment(client)
+    score_set = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
+    change_ownership(session, score_set["urn"], ScoreSetDbModel)
+
+    urn_with_space = score_set["urn"] + "   "
     search_payload = {"urn": urn_with_space}
     response = client.post("/api/v1/me/score-sets/search", json=search_payload)
     assert response.status_code == 200
@@ -1670,20 +1595,13 @@ def test_search_others_private_score_sets_urn_with_space_match(
 
 
 def test_search_public_score_sets_no_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client, {"title": "Experiment 1"})
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_1["urn"],
-        data_files / "scores.csv",
-        update={"title": "Test Score Set"},
-    )
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        score_set_response = client.post(f"/api/v1/score-sets/{score_set_1_1['urn']}/publish")
-        assert score_set_response.status_code == 200
-        queue.assert_called_once()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
     search_payload = {"text": "fnord"}
     response = client.post("/api/v1/score-sets/search", json=search_payload)
@@ -1692,38 +1610,29 @@ def test_search_public_score_sets_no_match(session, data_provider, client, setup
 
 
 def test_search_public_score_sets_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client, {"title": "Experiment 1"})
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_1["urn"],
-        data_files / "scores.csv",
-        update={"title": "Test Fnord Score Set"},
-    )
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"], update={"title": "Test Fnord Score Set"})
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        score_set_response = client.post(f"/api/v1/score-sets/{score_set_1_1['urn']}/publish")
-        assert score_set_response.status_code == 200
-        queue.assert_called_once()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
     search_payload = {"text": "fnord"}
     response = client.post("/api/v1/score-sets/search", json=search_payload)
     assert response.status_code == 200
     assert len(response.json()) == 1
-    assert response.json()[0]["title"] == score_set_1_1["title"]
+    assert response.json()[0]["title"] == score_set["title"]
 
 
 def test_search_public_score_sets_urn_with_space_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client)
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment_1["urn"], data_files / "scores.csv"
-    )
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        score_set_response = client.post(f"/api/v1/score-sets/{score_set_1_1['urn']}/publish")
-        published_score_set = score_set_response.json()
-        assert score_set_response.status_code == 200
-        queue.assert_called_once()
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
+
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
     urn_with_space = published_score_set["urn"] + "   "
     search_payload = {"urn": urn_with_space}
@@ -1734,23 +1643,16 @@ def test_search_public_score_sets_urn_with_space_match(session, data_provider, c
 
 
 def test_search_others_public_score_sets_no_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client, {"title": "Experiment 1"})
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_1["urn"],
-        data_files / "scores.csv",
-        update={"title": "Test Score Set"},
-    )
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        score_set_response = client.post(f"/api/v1/score-sets/{score_set_1_1['urn']}/publish")
-        assert score_set_response.status_code == 200
-        queue.assert_called_once()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
-    publish_score_set = score_set_response.json()
-    change_ownership(session, publish_score_set["urn"], ScoreSetDbModel)
+    change_ownership(session, published_score_set["urn"], ScoreSetDbModel)
+
     search_payload = {"text": "fnord"}
     response = client.post("/api/v1/score-sets/search", json=search_payload)
     assert response.status_code == 200
@@ -1758,65 +1660,52 @@ def test_search_others_public_score_sets_no_match(session, data_provider, client
 
 
 def test_search_others_public_score_sets_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client, {"title": "Experiment 1"})
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_1["urn"],
-        data_files / "scores.csv",
-        update={"title": "Test Fnord Score Set"},
-    )
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"], update={"title": "Test Fnord Score Set"})
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        score_set_response = client.post(f"/api/v1/score-sets/{score_set_1_1['urn']}/publish")
-        assert score_set_response.status_code == 200
-        queue.assert_called_once()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
-    publish_score_set = score_set_response.json()
-    change_ownership(session, publish_score_set["urn"], ScoreSetDbModel)
-    assert session.query(ScoreSetDbModel).filter_by(urn=publish_score_set["urn"]).one()
+    change_ownership(session, published_score_set["urn"], ScoreSetDbModel)
+    assert session.query(ScoreSetDbModel).filter_by(urn=published_score_set["urn"]).one()
+
     search_payload = {"text": "fnord"}
     response = client.post("/api/v1/score-sets/search", json=search_payload)
     assert response.status_code == 200
     assert len(response.json()) == 1
-    assert response.json()[0]["title"] == publish_score_set["title"]
+    assert response.json()[0]["title"] == published_score_set["title"]
 
 
 def test_search_others_public_score_sets_urn_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client)
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment_1["urn"], data_files / "scores.csv"
-    )
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        score_set_response = client.post(f"/api/v1/score-sets/{score_set_1_1['urn']}/publish")
-        assert score_set_response.status_code == 200
-        queue.assert_called_once()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
-    publish_score_set = score_set_response.json()
-    change_ownership(session, publish_score_set["urn"], ScoreSetDbModel)
-    search_payload = {"urn": score_set_1_1["urn"]}
+    change_ownership(session, published_score_set["urn"], ScoreSetDbModel)
+    search_payload = {"urn": score_set["urn"]}
     response = client.post("/api/v1/score-sets/search", json=search_payload)
     assert response.status_code == 200
     assert len(response.json()) == 1
-    assert response.json()[0]["urn"] == publish_score_set["urn"]
+    assert response.json()[0]["urn"] == published_score_set["urn"]
 
 
 def test_search_others_public_score_sets_urn_with_space_match(
     session, data_provider, client, setup_router_db, data_files
 ):
-    experiment_1 = create_experiment(client)
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment_1["urn"], data_files / "scores.csv"
-    )
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        score_set_response = client.post(f"/api/v1/score-sets/{score_set_1_1['urn']}/publish")
-        assert score_set_response.status_code == 200
-        queue.assert_called_once()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
-    published_score_set = score_set_response.json()
     change_ownership(session, published_score_set["urn"], ScoreSetDbModel)
     urn_with_space = published_score_set["urn"] + "   "
     search_payload = {"urn": urn_with_space}
@@ -1829,46 +1718,41 @@ def test_search_others_public_score_sets_urn_with_space_match(
 def test_search_private_score_sets_not_showing_public_score_set(
     session, data_provider, client, setup_router_db, data_files
 ):
-    experiment_1 = create_experiment(client)
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment_1["urn"], data_files / "scores.csv"
-    )
-    score_set_1_2 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment_1["urn"], data_files / "scores.csv"
-    )
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set_1 = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set_1 = mock_worker_variant_insertion(client, session, data_provider, score_set_1, data_files / "scores.csv")
+    score_set_2 = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 2"})
+    score_set_2 = mock_worker_variant_insertion(client, session, data_provider, score_set_2, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        score_set_response = client.post(f"/api/v1/score-sets/{score_set_1_1['urn']}/publish")
-        assert score_set_response.status_code == 200
-        queue.assert_called_once()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        publish_score_set(client, score_set_1["urn"])
+        worker_queue.assert_called_once()
 
     search_payload = {"published": False}
     response = client.post("/api/v1/score-sets/search", json=search_payload)
     assert response.status_code == 200
     assert len(response.json()) == 1
-    assert response.json()[0]["urn"] == score_set_1_2["urn"]
+    assert response.json()[0]["urn"] == score_set_2["urn"]
 
 
 def test_search_public_score_sets_not_showing_private_score_set(
     session, data_provider, client, setup_router_db, data_files
 ):
-    experiment_1 = create_experiment(client)
-    score_set_1_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment_1["urn"], data_files / "scores.csv"
-    )
-    create_seq_score_set_with_variants(client, session, data_provider, experiment_1["urn"], data_files / "scores.csv")
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set_1 = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 1"})
+    score_set_1 = mock_worker_variant_insertion(client, session, data_provider, score_set_1, data_files / "scores.csv")
+    score_set_2 = create_seq_score_set(client, experiment["urn"], update={"title": "Score Set 2"})
+    score_set_2 = mock_worker_variant_insertion(client, session, data_provider, score_set_2, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        score_set_response = client.post(f"/api/v1/score-sets/{score_set_1_1['urn']}/publish")
-        assert score_set_response.status_code == 200
-        queue.assert_called_once()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set_1 = publish_score_set(client, score_set_1["urn"])
+        worker_queue.assert_called_once()
 
-    published_score_set = score_set_response.json()
     search_payload = {"published": True}
     response = client.post("/api/v1/score-sets/search", json=search_payload)
     assert response.status_code == 200
     assert len(response.json()) == 1
-    assert response.json()[0]["urn"] == published_score_set["urn"]
+    assert response.json()[0]["urn"] == published_score_set_1["urn"]
 
 
 ########################################################################################################################
@@ -1880,9 +1764,8 @@ def test_anonymous_cannot_delete_other_users_private_scoreset(
     session, data_provider, client, setup_router_db, data_files, anonymous_app_overrides
 ):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
     with DependencyOverrider(anonymous_app_overrides):
         response = client.delete(f"/api/v1/score-sets/{score_set['urn']}")
@@ -1895,18 +1778,15 @@ def test_anonymous_cannot_delete_other_users_published_scoreset(
     session, data_provider, client, setup_router_db, data_files, anonymous_app_overrides
 ):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
-        assert response.status_code == 200
-        queue.assert_called_once()
-    response_data = response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
     with DependencyOverrider(anonymous_app_overrides):
-        del_response = client.delete(f"/api/v1/score-sets/{response_data['urn']}")
+        del_response = client.delete(f"/api/v1/score-sets/{published_score_set['urn']}")
 
     assert del_response.status_code == 401
     del_response_data = del_response.json()
@@ -1915,9 +1795,8 @@ def test_anonymous_cannot_delete_other_users_published_scoreset(
 
 def test_can_delete_own_private_scoreset(session, data_provider, client, setup_router_db, data_files):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
     response = client.delete(f"/api/v1/score-sets/{score_set['urn']}")
 
@@ -1926,30 +1805,26 @@ def test_can_delete_own_private_scoreset(session, data_provider, client, setup_r
 
 def test_cannot_delete_own_published_scoreset(session, data_provider, client, setup_router_db, data_files):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
-        assert response.status_code == 200
-        queue.assert_called_once()
-    response_data = response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
-    del_response = client.delete(f"/api/v1/score-sets/{response_data['urn']}")
+    del_response = client.delete(f"/api/v1/score-sets/{published_score_set['urn']}")
 
     assert del_response.status_code == 403
     del_response_data = del_response.json()
-    assert f"insufficient permissions for URN '{response_data['urn']}'" in del_response_data["detail"]
+    assert f"insufficient permissions for URN '{published_score_set['urn']}'" in del_response_data["detail"]
 
 
 def test_contributor_can_delete_other_users_private_scoreset(
     session, data_provider, client, setup_router_db, data_files, admin_app_overrides
 ):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
     change_ownership(session, score_set["urn"], ScoreSetDbModel)
     add_contributor(
         session,
@@ -1969,9 +1844,8 @@ def test_admin_can_delete_other_users_private_scoreset(
     session, data_provider, client, setup_router_db, data_files, admin_app_overrides
 ):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
     with DependencyOverrider(admin_app_overrides):
         response = client.delete(f"/api/v1/score-sets/{score_set['urn']}")
@@ -1983,20 +1857,16 @@ def test_admin_can_delete_other_users_published_scoreset(
     session, data_provider, client, setup_router_db, data_files, admin_app_overrides
 ):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
-        assert response.status_code == 200
-        queue.assert_called_once()
-    response_data = response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
     with DependencyOverrider(admin_app_overrides):
-        del_response = client.delete(f"/api/v1/score-sets/{response_data['urn']}")
-
-    assert del_response.status_code == 200
+        del_response = client.delete(f"/api/v1/score-sets/{published_score_set['urn']}")
+        assert del_response.status_code == 200
 
 
 ########################################################################################################################
@@ -2026,33 +1896,32 @@ def test_cannot_add_score_set_to_others_private_experiment(session, client, setu
 
 def test_can_add_score_set_to_own_public_experiment(session, data_provider, client, setup_router_db, data_files):
     experiment = create_experiment(client)
-    score_set_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set_1 = create_seq_score_set(client, experiment["urn"])
+    score_set_1 = mock_worker_variant_insertion(client, session, data_provider, score_set_1, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        pub_score_set_1 = (client.post(f"/api/v1/score-sets/{score_set_1['urn']}/publish")).json()
-        queue.assert_called_once()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set_1 = publish_score_set(client, score_set_1["urn"])
+        worker_queue.assert_called_once()
 
     score_set_2 = deepcopy(TEST_MINIMAL_SEQ_SCORESET)
-    score_set_2["experimentUrn"] = pub_score_set_1["experiment"]["urn"]
+    score_set_2["experimentUrn"] = published_score_set_1["experiment"]["urn"]
     response = client.post("/api/v1/score-sets/", json=score_set_2)
     assert response.status_code == 200
 
 
 def test_can_add_score_set_to_others_public_experiment(session, data_provider, client, setup_router_db, data_files):
     experiment = create_experiment(client)
-    score_set_1 = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set_1 = create_seq_score_set(client, experiment["urn"])
+    score_set_1 = mock_worker_variant_insertion(client, session, data_provider, score_set_1, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        pub_score_set_1 = (client.post(f"/api/v1/score-sets/{score_set_1['urn']}/publish")).json()
-        queue.assert_called_once()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set_1["urn"])
+        worker_queue.assert_called_once()
 
-    change_ownership(session, pub_score_set_1["experiment"]["urn"], ExperimentDbModel)
+    published_experiment_urn = published_score_set["experiment"]["urn"]
+    change_ownership(session, published_experiment_urn, ExperimentDbModel)
     score_set_2 = deepcopy(TEST_MINIMAL_SEQ_SCORESET)
-    score_set_2["experimentUrn"] = pub_score_set_1["experiment"]["urn"]
+    score_set_2["experimentUrn"] = published_experiment_urn
     response = client.post("/api/v1/score-sets/", json=score_set_2)
     assert response.status_code == 200
 
@@ -2078,25 +1947,25 @@ def test_contributor_can_add_score_set_to_others_public_experiment(
     session, data_provider, client, setup_router_db, data_files
 ):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        published_score_set = (client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")).json()
-        queue.assert_called_once()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
-    change_ownership(session, published_score_set["experiment"]["urn"], ExperimentDbModel)
+    published_experiment_urn = published_score_set["experiment"]["urn"]
+    change_ownership(session, published_experiment_urn, ExperimentDbModel)
     add_contributor(
         session,
-        published_score_set["experiment"]["urn"],
+        published_experiment_urn,
         ExperimentDbModel,
         TEST_USER["username"],
         TEST_USER["first_name"],
         TEST_USER["last_name"],
     )
     score_set_post_payload = deepcopy(TEST_MINIMAL_SEQ_SCORESET)
-    score_set_post_payload["experimentUrn"] = published_score_set["experiment"]["urn"]
+    score_set_post_payload["experimentUrn"] = published_experiment_urn
     response = client.post("/api/v1/score-sets/", json=score_set_post_payload)
     assert response.status_code == 200
 
@@ -2138,15 +2007,13 @@ def test_can_modify_metadata_for_score_set_with_inactive_license(session, client
 
 def test_create_superseding_score_set(session, data_provider, client, setup_router_db, data_files):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        publish_score_set_response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
-        assert publish_score_set_response.status_code == 200
-        queue.assert_called_once()
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    published_score_set = publish_score_set_response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
+
     score_set_post_payload = deepcopy(TEST_MINIMAL_SEQ_SCORESET)
     score_set_post_payload["experimentUrn"] = published_score_set["experiment"]["urn"]
     score_set_post_payload["supersededScoreSetUrn"] = published_score_set["urn"]
@@ -2156,15 +2023,15 @@ def test_create_superseding_score_set(session, data_provider, client, setup_rout
 
 def test_can_view_unpublished_superseding_score_set(session, data_provider, client, setup_router_db, data_files):
     experiment = create_experiment(client)
-    unpublished_score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
+    unpublished_score_set = create_seq_score_set(client, experiment["urn"])
+    unpublished_score_set = mock_worker_variant_insertion(
+        client, session, data_provider, unpublished_score_set, data_files / "scores.csv"
     )
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        publish_score_set_response = client.post(f"/api/v1/score-sets/{unpublished_score_set['urn']}/publish")
-        assert publish_score_set_response.status_code == 200
-        queue.assert_called_once()
 
-    published_score_set = publish_score_set_response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, unpublished_score_set["urn"])
+        worker_queue.assert_called_once()
+
     score_set_post_payload = deepcopy(TEST_MINIMAL_SEQ_SCORESET)
     score_set_post_payload["experimentUrn"] = published_score_set["experiment"]["urn"]
     score_set_post_payload["supersededScoreSetUrn"] = published_score_set["urn"]
@@ -2182,15 +2049,14 @@ def test_cannot_view_others_unpublished_superseding_score_set(
     session, data_provider, client, setup_router_db, data_files
 ):
     experiment = create_experiment(client)
-    unpublished_score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
+    unpublished_score_set = create_seq_score_set(client, experiment["urn"])
+    unpublished_score_set = mock_worker_variant_insertion(
+        client, session, data_provider, unpublished_score_set, data_files / "scores.csv"
     )
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        publish_score_set_response = client.post(f"/api/v1/score-sets/{unpublished_score_set['urn']}/publish")
-        assert publish_score_set_response.status_code == 200
-        queue.assert_called_once()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, unpublished_score_set["urn"])
+        worker_queue.assert_called_once()
 
-    published_score_set = publish_score_set_response.json()
     score_set_post_payload = deepcopy(TEST_MINIMAL_SEQ_SCORESET)
     score_set_post_payload["experimentUrn"] = published_score_set["experiment"]["urn"]
     score_set_post_payload["supersededScoreSetUrn"] = published_score_set["urn"]
@@ -2208,30 +2074,24 @@ def test_cannot_view_others_unpublished_superseding_score_set(
 
 def test_can_view_others_published_superseding_score_set(session, data_provider, client, setup_router_db, data_files):
     experiment = create_experiment(client)
-    unpublished_score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
+    unpublished_score_set = create_seq_score_set(client, experiment["urn"])
+    unpublished_score_set = mock_worker_variant_insertion(
+        client, session, data_provider, unpublished_score_set, data_files / "scores.csv"
     )
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        publish_score_set_response = client.post(f"/api/v1/score-sets/{unpublished_score_set['urn']}/publish")
-        assert publish_score_set_response.status_code == 200
-        queue.assert_called_once()
-    published_score_set = publish_score_set_response.json()
 
-    superseding_score_set = create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        published_score_set["experiment"]["urn"],
-        data_files / "scores.csv",
-        update={"supersededScoreSetUrn": published_score_set["urn"]},
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, unpublished_score_set["urn"])
+        worker_queue.assert_called_once()
+
+    superseding_score_set = create_seq_score_set(
+        client, published_score_set["experiment"]["urn"], update={"supersededScoreSetUrn": published_score_set["urn"]}
     )
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        published_superseding_score_set_response = client.post(
-            f"/api/v1/score-sets/{superseding_score_set['urn']}/publish"
-        )
-        assert publish_score_set_response.status_code == 200
-        queue.assert_called_once()
-    published_superseding_score_set = published_superseding_score_set_response.json()
+    superseding_score_set = mock_worker_variant_insertion(
+        client, session, data_provider, superseding_score_set, data_files / "scores.csv"
+    )
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_superseding_score_set = publish_score_set(client, superseding_score_set["urn"])
+        worker_queue.assert_called_once()
 
     change_ownership(session, published_superseding_score_set["urn"], ScoreSetDbModel)
 
@@ -2248,14 +2108,14 @@ def test_show_correct_score_set_version_with_superseded_score_set_to_its_owner(
     session, data_provider, client, setup_router_db, data_files
 ):
     experiment = create_experiment(client)
-    unpublished_score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
+    unpublished_score_set = create_seq_score_set(client, experiment["urn"])
+    unpublished_score_set = mock_worker_variant_insertion(
+        client, session, data_provider, unpublished_score_set, data_files / "scores.csv"
     )
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        publish_score_set_response = client.post(f"/api/v1/score-sets/{unpublished_score_set['urn']}/publish")
-        assert publish_score_set_response.status_code == 200
-        queue.assert_called_once()
-    published_score_set = publish_score_set_response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, unpublished_score_set["urn"])
+        worker_queue.assert_called_once()
+
     score_set_post_payload = deepcopy(TEST_MINIMAL_SEQ_SCORESET)
     score_set_post_payload["experimentUrn"] = published_score_set["experiment"]["urn"]
     score_set_post_payload["supersededScoreSetUrn"] = published_score_set["urn"]
@@ -2266,6 +2126,11 @@ def test_show_correct_score_set_version_with_superseded_score_set_to_its_owner(
     score_set = score_set_response.json()
     assert score_set_response.status_code == 200
     assert score_set["urn"] == superseding_score_set["urn"]
+
+
+########################################################################################################################
+# Score Calibrations
+########################################################################################################################
 
 
 def test_anonymous_user_cannot_add_score_calibrations_to_score_set(client, setup_router_db, anonymous_app_overrides):
@@ -2367,18 +2232,15 @@ def test_upload_a_non_utf8_file(session, client, setup_router_db, data_files):
 # Test file doesn't have hgvs_splice so its values are all NA.
 def test_download_scores_file(session, data_provider, client, setup_router_db, data_files):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client, session, data_provider, experiment["urn"], data_files / "scores.csv"
-    )
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        publish_score_set_response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
-        assert publish_score_set_response.status_code == 200
-        queue.assert_called_once()
-    publish_score_set = publish_score_set_response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
     download_scores_csv_response = client.get(
-        f"/api/v1/score-sets/{publish_score_set['urn']}/scores?drop_na_columns=true"
+        f"/api/v1/score-sets/{published_score_set['urn']}/scores?drop_na_columns=true"
     )
     assert download_scores_csv_response.status_code == 200
     download_scores_csv = download_scores_csv_response.text
@@ -2391,22 +2253,16 @@ def test_download_scores_file(session, data_provider, client, setup_router_db, d
 
 def test_download_counts_file(session, data_provider, client, setup_router_db, data_files):
     experiment = create_experiment(client)
-    score_set = create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment["urn"],
-        scores_csv_path=data_files / "scores.csv",
-        counts_csv_path=data_files / "counts.csv",
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(
+        client, session, data_provider, score_set, data_files / "scores.csv", data_files / "counts.csv"
     )
-    with patch.object(ArqRedis, "enqueue_job", return_value=None) as queue:
-        publish_score_set_response = client.post(f"/api/v1/score-sets/{score_set['urn']}/publish")
-        assert publish_score_set_response.status_code == 200
-        queue.assert_called_once()
-    publish_score_set = publish_score_set_response.json()
+    with patch.object(arq.ArqRedis, "enqueue_job", return_value=None) as worker_queue:
+        published_score_set = publish_score_set(client, score_set["urn"])
+        worker_queue.assert_called_once()
 
     download_counts_csv_response = client.get(
-        f"/api/v1/score-sets/{publish_score_set['urn']}/counts?drop_na_columns=true"
+        f"/api/v1/score-sets/{published_score_set['urn']}/counts?drop_na_columns=true"
     )
     assert download_counts_csv_response.status_code == 200
     download_counts_csv = download_counts_csv_response.text
