+++ tests/routers/test_target_gene.py
@@ -1,21 +1,22 @@
+# ruff: noqa: E402
+import pytest
+
+arq = pytest.importorskip("arq")
+cdot = pytest.importorskip("cdot")
+fastapi = pytest.importorskip("fastapi")
+
 from mavedb.models.score_set import ScoreSet as ScoreSetDbModel
-from tests.helpers.util import (
-    change_ownership,
-    create_experiment,
-    create_seq_score_set_with_variants,
-)
+
+from tests.helpers.util.experiment import create_experiment
+from tests.helpers.util.user import change_ownership
+from tests.helpers.util.score_set import create_seq_score_set
+from tests.helpers.util.variant import mock_worker_variant_insertion
 
 
 def test_search_my_target_genes_no_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client, {"title": "Experiment 1"})
-    create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_1["urn"],
-        data_files / "scores.csv",
-        update={"title": "Test Score Set"},
-    )
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
     search_payload = {"text": "NONEXISTENT"}
     response = client.post("/api/v1/me/target-genes/search", json=search_payload)
@@ -24,15 +25,9 @@ def test_search_my_target_genes_no_match(session, data_provider, client, setup_r
 
 
 def test_search_my_target_genes_no_match_on_other_user(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client, {"title": "Experiment 1"})
-    score_set = create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_1["urn"],
-        data_files / "scores.csv",
-        update={"title": "Test Score Set"},
-    )
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
     change_ownership(session, score_set["urn"], ScoreSetDbModel)
 
     search_payload = {"text": "TEST1"}
@@ -42,15 +37,9 @@ def test_search_my_target_genes_no_match_on_other_user(session, data_provider, c
 
 
 def test_search_my_target_genes_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client, {"title": "Experiment 1"})
-    create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_1["urn"],
-        data_files / "scores.csv",
-        update={"title": "Test Score Set"},
-    )
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
     search_payload = {"text": "TEST1"}
     response = client.post("/api/v1/me/target-genes/search", json=search_payload)
@@ -60,15 +49,9 @@ def test_search_my_target_genes_match(session, data_provider, client, setup_rout
 
 
 def test_search_target_genes_no_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client, {"title": "Experiment 1"})
-    create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_1["urn"],
-        data_files / "scores.csv",
-        update={"title": "Test Score Set"},
-    )
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
     search_payload = {"text": "NONEXISTENT"}
     response = client.post("/api/v1/target-genes/search", json=search_payload)
@@ -77,15 +60,9 @@ def test_search_target_genes_no_match(session, data_provider, client, setup_rout
 
 
 def test_search_target_genes_match_on_other_user(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client, {"title": "Experiment 1"})
-    score_set = create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_1["urn"],
-        data_files / "scores.csv",
-        update={"title": "Test Score Set"},
-    )
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
     change_ownership(session, score_set["urn"], ScoreSetDbModel)
 
     search_payload = {"text": "TEST1"}
@@ -96,15 +73,9 @@ def test_search_target_genes_match_on_other_user(session, data_provider, client,
 
 
 def test_search_target_genes_match(session, data_provider, client, setup_router_db, data_files):
-    experiment_1 = create_experiment(client, {"title": "Experiment 1"})
-    create_seq_score_set_with_variants(
-        client,
-        session,
-        data_provider,
-        experiment_1["urn"],
-        data_files / "scores.csv",
-        update={"title": "Test Score Set"},
-    )
+    experiment = create_experiment(client, {"title": "Experiment 1"})
+    score_set = create_seq_score_set(client, experiment["urn"])
+    score_set = mock_worker_variant_insertion(client, session, data_provider, score_set, data_files / "scores.csv")
 
     search_payload = {"text": "TEST1"}
     response = client.post("/api/v1/target-genes/search", json=search_payload)
