+++ tests/view_models/test_experiment.py
@@ -1,5 +1,4 @@
 import pytest
-from fastapi.encoders import jsonable_encoder
 
 from mavedb.view_models.experiment import ExperimentCreate
 from tests.helpers.constants import TEST_MINIMAL_EXPERIMENT
@@ -7,7 +6,7 @@ from tests.helpers.constants import TEST_MINIMAL_EXPERIMENT
 
 # Test valid experiment
 def test_create_experiment():
-    experiment = ExperimentCreate(**jsonable_encoder(TEST_MINIMAL_EXPERIMENT))
+    experiment = ExperimentCreate(**TEST_MINIMAL_EXPERIMENT)
     assert experiment.title == "Test Experiment Title"
     assert experiment.short_description == "Test experiment"
     assert experiment.abstract_text == "Abstract"
@@ -16,9 +15,9 @@ def test_create_experiment():
 
 def test_cannot_create_experiment_without_a_title():
     experiment = TEST_MINIMAL_EXPERIMENT.copy()
-    invalid_experiment = jsonable_encoder(experiment, exclude={"title"})
+    experiment.pop("title")
     with pytest.raises(ValueError) as exc_info:
-        ExperimentCreate(**invalid_experiment)
+        ExperimentCreate(**experiment)
 
     assert "field required" in str(exc_info.value)
     assert "title" in str(exc_info.value)
@@ -26,11 +25,10 @@ def test_cannot_create_experiment_without_a_title():
 
 def test_cannot_create_experiment_with_a_space_title():
     experiment = TEST_MINIMAL_EXPERIMENT.copy()
-    invalid_experiment = jsonable_encoder(experiment, exclude={"title"})
-    invalid_experiment["title"] = " "
+    experiment["title"] = " "
 
     with pytest.raises(ValueError) as exc_info:
-        ExperimentCreate(**invalid_experiment)
+        ExperimentCreate(**experiment)
 
     assert "This field is required and cannot be empty." in str(exc_info.value)
     assert "title" in str(exc_info.value)
@@ -38,11 +36,10 @@ def test_cannot_create_experiment_with_a_space_title():
 
 def test_cannot_create_experiment_with_an_empty_title():
     experiment = TEST_MINIMAL_EXPERIMENT.copy()
-    invalid_experiment = jsonable_encoder(experiment, exclude={"title"})
-    invalid_experiment["title"] = ""
+    experiment["title"] = ""
 
     with pytest.raises(ValueError) as exc_info:
-        ExperimentCreate(**invalid_experiment)
+        ExperimentCreate(**experiment)
 
     assert "none is not an allowed value" in str(exc_info.value)
     assert "title" in str(exc_info.value)
@@ -50,10 +47,10 @@ def test_cannot_create_experiment_with_an_empty_title():
 
 def test_cannot_create_experiment_without_a_short_description():
     experiment = TEST_MINIMAL_EXPERIMENT.copy()
-    invalid_experiment = jsonable_encoder(experiment, exclude={"shortDescription"})
+    experiment.pop("shortDescription")
 
     with pytest.raises(ValueError) as exc_info:
-        ExperimentCreate(**invalid_experiment)
+        ExperimentCreate(**experiment)
 
     assert "field required" in str(exc_info.value)
     assert "shortDescription" in str(exc_info.value)
@@ -61,11 +58,10 @@ def test_cannot_create_experiment_without_a_short_description():
 
 def test_cannot_create_experiment_with_a_space_short_description():
     experiment = TEST_MINIMAL_EXPERIMENT.copy()
-    invalid_experiment = jsonable_encoder(experiment, exclude={"shortDescription"})
-    invalid_experiment["shortDescription"] = "  "
+    experiment["shortDescription"] = "  "
 
     with pytest.raises(ValueError) as exc_info:
-        ExperimentCreate(**invalid_experiment)
+        ExperimentCreate(**experiment)
 
     assert "This field is required and cannot be empty." in str(exc_info.value)
     assert "shortDescription" in str(exc_info.value)
@@ -73,11 +69,10 @@ def test_cannot_create_experiment_with_a_space_short_description():
 
 def test_cannot_create_experiment_with_an_empty_short_description():
     experiment = TEST_MINIMAL_EXPERIMENT.copy()
-    invalid_experiment = jsonable_encoder(experiment, exclude={"shortDescription"})
-    invalid_experiment["shortDescription"] = ""
+    experiment["shortDescription"] = ""
 
     with pytest.raises(ValueError) as exc_info:
-        ExperimentCreate(**invalid_experiment)
+        ExperimentCreate(**experiment)
 
     assert "none is not an allowed value" in str(exc_info.value)
     assert "shortDescription" in str(exc_info.value)
@@ -85,10 +80,10 @@ def test_cannot_create_experiment_with_an_empty_short_description():
 
 def test_cannot_create_experiment_without_an_abstract():
     experiment = TEST_MINIMAL_EXPERIMENT.copy()
-    invalid_experiment = jsonable_encoder(experiment, exclude={"abstractText"})
+    experiment.pop("abstractText")
 
     with pytest.raises(ValueError) as exc_info:
-        ExperimentCreate(**invalid_experiment)
+        ExperimentCreate(**experiment)
 
     assert "field required" in str(exc_info.value)
     assert "abstractText" in str(exc_info.value)
@@ -96,11 +91,10 @@ def test_cannot_create_experiment_without_an_abstract():
 
 def test_cannot_create_experiment_with_a_space_abstract():
     experiment = TEST_MINIMAL_EXPERIMENT.copy()
-    invalid_experiment = jsonable_encoder(experiment, exclude={"abstractText"})
-    invalid_experiment["abstractText"] = "  "
+    experiment["abstractText"] = "  "
 
     with pytest.raises(ValueError) as exc_info:
-        ExperimentCreate(**invalid_experiment)
+        ExperimentCreate(**experiment)
 
     assert "This field is required and cannot be empty." in str(exc_info.value)
     assert "abstractText" in str(exc_info.value)
@@ -108,11 +102,10 @@ def test_cannot_create_experiment_with_a_space_abstract():
 
 def test_cannot_create_experiment_with_an_empty_abstract():
     experiment = TEST_MINIMAL_EXPERIMENT.copy()
-    invalid_experiment = jsonable_encoder(experiment, exclude={"abstractText"})
-    invalid_experiment["abstractText"] = ""
+    experiment["abstractText"] = ""
 
     with pytest.raises(ValueError) as exc_info:
-        ExperimentCreate(**invalid_experiment)
+        ExperimentCreate(**experiment)
 
     assert "none is not an allowed value" in str(exc_info.value)
     assert "abstractText" in str(exc_info.value)
@@ -120,10 +113,10 @@ def test_cannot_create_experiment_with_an_empty_abstract():
 
 def test_cannot_create_experiment_without_a_method():
     experiment = TEST_MINIMAL_EXPERIMENT.copy()
-    invalid_experiment = jsonable_encoder(experiment, exclude={"methodText"})
+    experiment.pop("methodText")
 
     with pytest.raises(ValueError) as exc_info:
-        ExperimentCreate(**invalid_experiment)
+        ExperimentCreate(**experiment)
 
     assert "field required" in str(exc_info.value)
     assert "methodText" in str(exc_info.value)
@@ -131,11 +124,10 @@ def test_cannot_create_experiment_without_a_method():
 
 def test_cannot_create_experiment_with_a_space_method():
     experiment = TEST_MINIMAL_EXPERIMENT.copy()
-    invalid_experiment = jsonable_encoder(experiment, exclude={"methodText"})
-    invalid_experiment["methodText"] = "  "
+    experiment["methodText"] = "  "
 
     with pytest.raises(ValueError) as exc_info:
-        ExperimentCreate(**invalid_experiment)
+        ExperimentCreate(**experiment)
 
     assert "This field is required and cannot be empty." in str(exc_info.value)
     assert "methodText" in str(exc_info.value)
@@ -143,11 +135,10 @@ def test_cannot_create_experiment_with_a_space_method():
 
 def test_cannot_create_experiment_with_an_empty_method():
     experiment = TEST_MINIMAL_EXPERIMENT.copy()
-    invalid_experiment = jsonable_encoder(experiment, exclude={"methodText"})
-    invalid_experiment["methodText"] = ""
+    experiment["methodText"] = ""
 
     with pytest.raises(ValueError) as exc_info:
-        ExperimentCreate(**invalid_experiment)
+        ExperimentCreate(**experiment)
 
     assert "none is not an allowed value" in str(exc_info.value)
     assert "methodText" in str(exc_info.value)
