+++ tests/view_models/test_score_set.py
@@ -1,5 +1,4 @@
 import pytest
-from fastapi.encoders import jsonable_encoder
 
 from mavedb.view_models.publication_identifier import PublicationIdentifierCreate
 from mavedb.view_models.score_set import ScoreSetCreate, ScoreSetModify
@@ -9,24 +8,25 @@ from tests.helpers.constants import TEST_MINIMAL_SEQ_SCORESET
 
 def test_cannot_create_score_set_without_a_target():
     score_set_test = TEST_MINIMAL_SEQ_SCORESET.copy()
+    score_set_test.pop("targetGenes")
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetModify(**jsonable_encoder(score_set_test, exclude={"targetGenes"}), target_genes=[])
+        ScoreSetModify(**score_set_test, target_genes=[])
 
     assert "Score sets should define at least one target." in str(exc_info.value)
 
 
 def test_cannot_create_score_set_with_multiple_primary_publications():
     score_set_test = TEST_MINIMAL_SEQ_SCORESET.copy()
+    target_genes = score_set_test.pop("targetGenes")
 
     identifier_one = PublicationIdentifierCreate(identifier="2019.12.12.207222")
     identifier_two = PublicationIdentifierCreate(identifier="2019.12.12.20733333")
 
     with pytest.raises(ValueError) as exc_info:
         ScoreSetModify(
-            **jsonable_encoder(score_set_test),
-            exclude={"targetGenes"},
-            target_genes=[TargetGeneCreate(**jsonable_encoder(target)) for target in score_set_test["targetGenes"]],
+            **score_set_test,
+            target_genes=[TargetGeneCreate(**target) for target in target_genes],
             primary_publication_identifiers=[identifier_one, identifier_two],
         )
 
@@ -36,12 +36,13 @@ def test_cannot_create_score_set_with_multiple_primary_publications():
 def test_cannot_create_score_set_without_target_gene_labels_when_multiple_targets_exist():
     score_set_test = TEST_MINIMAL_SEQ_SCORESET.copy()
 
-    target_gene_one = TargetGeneCreate(**jsonable_encoder(score_set_test["targetGenes"][0]))
-    target_gene_two = TargetGeneCreate(**jsonable_encoder(score_set_test["targetGenes"][0]))
+    target_gene_one = TargetGeneCreate(**score_set_test["targetGenes"][0])
+    target_gene_two = TargetGeneCreate(**score_set_test["targetGenes"][0])
 
+    score_set_test.pop("targetGenes")
     with pytest.raises(ValueError) as exc_info:
         ScoreSetModify(
-            **jsonable_encoder(score_set_test, exclude={"targetGenes"}),
+            **score_set_test,
             target_genes=[target_gene_one, target_gene_two],
         )
 
@@ -51,16 +52,17 @@ def test_cannot_create_score_set_without_target_gene_labels_when_multiple_target
 def test_cannot_create_score_set_with_non_unique_target_labels():
     score_set_test = TEST_MINIMAL_SEQ_SCORESET.copy()
 
-    target_gene_one = TargetGeneCreate(**jsonable_encoder(score_set_test["targetGenes"][0]))
-    target_gene_two = TargetGeneCreate(**jsonable_encoder(score_set_test["targetGenes"][0]))
+    target_gene_one = TargetGeneCreate(**score_set_test["targetGenes"][0])
+    target_gene_two = TargetGeneCreate(**score_set_test["targetGenes"][0])
 
     non_unique = "BRCA1"
     target_gene_one.target_sequence.label = non_unique
     target_gene_two.target_sequence.label = non_unique
 
+    score_set_test.pop("targetGenes")
     with pytest.raises(ValueError) as exc_info:
         ScoreSetModify(
-            **jsonable_encoder(score_set_test, exclude={"targetGenes"}),
+            **score_set_test,
             target_genes=[target_gene_one, target_gene_two],
         )
 
@@ -69,9 +71,10 @@ def test_cannot_create_score_set_with_non_unique_target_labels():
 
 def test_cannot_create_score_set_without_a_title():
     score_set = TEST_MINIMAL_SEQ_SCORESET.copy()
-    invalid_score_set = jsonable_encoder(score_set, exclude={"title"})
+    score_set.pop("title")
+
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetCreate(**invalid_score_set)
+        ScoreSetCreate(**score_set)
 
     assert "field required" in str(exc_info.value)
     assert "title" in str(exc_info.value)
@@ -79,11 +82,10 @@ def test_cannot_create_score_set_without_a_title():
 
 def test_cannot_create_score_set_with_a_space_title():
     score_set = TEST_MINIMAL_SEQ_SCORESET.copy()
-    invalid_score_set = jsonable_encoder(score_set, exclude={"title"})
-    invalid_score_set["title"] = " "
+    score_set["title"] = " "
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetCreate(**invalid_score_set)
+        ScoreSetCreate(**score_set)
 
     assert "This field is required and cannot be empty." in str(exc_info.value)
     assert "title" in str(exc_info.value)
@@ -91,11 +93,10 @@ def test_cannot_create_score_set_with_a_space_title():
 
 def test_cannot_create_score_set_with_an_empty_title():
     score_set = TEST_MINIMAL_SEQ_SCORESET.copy()
-    invalid_score_set = jsonable_encoder(score_set, exclude={"title"})
-    invalid_score_set["title"] = ""
+    score_set["title"] = ""
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetCreate(**invalid_score_set)
+        ScoreSetCreate(**score_set)
 
     assert "none is not an allowed value" in str(exc_info.value)
     assert "title" in str(exc_info.value)
@@ -103,10 +104,10 @@ def test_cannot_create_score_set_with_an_empty_title():
 
 def test_cannot_create_score_set_without_a_short_description():
     score_set = TEST_MINIMAL_SEQ_SCORESET.copy()
-    invalid_score_set = jsonable_encoder(score_set, exclude={"shortDescription"})
+    score_set.pop("shortDescription")
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetCreate(**invalid_score_set)
+        ScoreSetCreate(**score_set)
 
     assert "field required" in str(exc_info.value)
     assert "shortDescription" in str(exc_info.value)
@@ -114,11 +115,10 @@ def test_cannot_create_score_set_without_a_short_description():
 
 def test_cannot_create_score_set_with_a_space_short_description():
     score_set = TEST_MINIMAL_SEQ_SCORESET.copy()
-    invalid_score_set = jsonable_encoder(score_set, exclude={"shortDescription"})
-    invalid_score_set["shortDescription"] = "  "
+    score_set["shortDescription"] = "  "
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetCreate(**invalid_score_set)
+        ScoreSetCreate(**score_set)
 
     assert "This field is required and cannot be empty." in str(exc_info.value)
     assert "shortDescription" in str(exc_info.value)
@@ -126,11 +126,10 @@ def test_cannot_create_score_set_with_a_space_short_description():
 
 def test_cannot_create_score_set_with_an_empty_short_description():
     score_set = TEST_MINIMAL_SEQ_SCORESET.copy()
-    invalid_score_set = jsonable_encoder(score_set, exclude={"shortDescription"})
-    invalid_score_set["shortDescription"] = ""
+    score_set["shortDescription"] = ""
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetCreate(**invalid_score_set)
+        ScoreSetCreate(**score_set)
 
     assert "none is not an allowed value" in str(exc_info.value)
     assert "shortDescription" in str(exc_info.value)
@@ -138,10 +137,10 @@ def test_cannot_create_score_set_with_an_empty_short_description():
 
 def test_cannot_create_score_set_without_an_abstract():
     score_set = TEST_MINIMAL_SEQ_SCORESET.copy()
-    invalid_score_set = jsonable_encoder(score_set, exclude={"abstractText"})
+    score_set.pop("abstractText")
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetCreate(**invalid_score_set)
+        ScoreSetCreate(**score_set)
 
     assert "field required" in str(exc_info.value)
     assert "abstractText" in str(exc_info.value)
@@ -149,11 +148,10 @@ def test_cannot_create_score_set_without_an_abstract():
 
 def test_cannot_create_score_set_with_a_space_abstract():
     score_set = TEST_MINIMAL_SEQ_SCORESET.copy()
-    invalid_score_set = jsonable_encoder(score_set, exclude={"abstractText"})
-    invalid_score_set["abstractText"] = "  "
+    score_set["abstractText"] = "  "
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetCreate(**invalid_score_set)
+        ScoreSetCreate(**score_set)
 
     assert "This field is required and cannot be empty." in str(exc_info.value)
     assert "abstractText" in str(exc_info.value)
@@ -161,11 +159,10 @@ def test_cannot_create_score_set_with_a_space_abstract():
 
 def test_cannot_create_score_set_with_an_empty_abstract():
     score_set = TEST_MINIMAL_SEQ_SCORESET.copy()
-    invalid_score_set = jsonable_encoder(score_set, exclude={"abstractText"})
-    invalid_score_set["abstractText"] = ""
+    score_set["abstractText"] = ""
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetCreate(**invalid_score_set)
+        ScoreSetCreate(**score_set)
 
     assert "none is not an allowed value" in str(exc_info.value)
     assert "abstractText" in str(exc_info.value)
@@ -173,10 +170,10 @@ def test_cannot_create_score_set_with_an_empty_abstract():
 
 def test_cannot_create_score_set_without_a_method():
     score_set = TEST_MINIMAL_SEQ_SCORESET.copy()
-    invalid_score_set = jsonable_encoder(score_set, exclude={"methodText"})
+    score_set.pop("methodText")
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetCreate(**invalid_score_set)
+        ScoreSetCreate(**score_set)
 
     assert "field required" in str(exc_info.value)
     assert "methodText" in str(exc_info.value)
@@ -184,11 +181,10 @@ def test_cannot_create_score_set_without_a_method():
 
 def test_cannot_create_score_set_with_a_space_method():
     score_set = TEST_MINIMAL_SEQ_SCORESET.copy()
-    invalid_score_set = jsonable_encoder(score_set, exclude={"methodText"})
-    invalid_score_set["methodText"] = "  "
+    score_set["methodText"] = "  "
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetCreate(**invalid_score_set)
+        ScoreSetCreate(**score_set)
 
     assert "This field is required and cannot be empty." in str(exc_info.value)
     assert "methodText" in str(exc_info.value)
@@ -196,11 +192,10 @@ def test_cannot_create_score_set_with_a_space_method():
 
 def test_cannot_create_score_set_with_an_empty_method():
     score_set = TEST_MINIMAL_SEQ_SCORESET.copy()
-    invalid_score_set = jsonable_encoder(score_set, exclude={"methodText"})
-    invalid_score_set["methodText"] = ""
+    score_set["methodText"] = ""
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetCreate(**invalid_score_set)
+        ScoreSetCreate(**score_set)
 
     assert "none is not an allowed value" in str(exc_info.value)
     assert "methodText" in str(exc_info.value)
@@ -217,7 +212,7 @@ def test_cannot_create_score_set_with_too_many_boundaries():
     }
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetModify(**jsonable_encoder(score_set_test))
+        ScoreSetModify(**score_set_test)
 
     assert "Only a lower and upper bound are allowed." in str(exc_info.value)
 
@@ -233,7 +228,7 @@ def test_cannot_create_score_set_with_overlapping_ranges():
     }
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetModify(**jsonable_encoder(score_set_test))
+        ScoreSetModify(**score_set_test)
 
     assert "Score ranges may not overlap; `range_1` overlaps with `range_2`" in str(exc_info.value)
 
@@ -249,7 +244,7 @@ def test_can_create_score_set_with_mixed_range_types():
         ],
     }
 
-    ScoreSetModify(**jsonable_encoder(score_set_test))
+    ScoreSetModify(**score_set_test)
 
 
 def test_can_create_score_set_with_adjacent_ranges():
@@ -262,7 +257,7 @@ def test_can_create_score_set_with_adjacent_ranges():
         ],
     }
 
-    ScoreSetModify(**jsonable_encoder(score_set_test))
+    ScoreSetModify(**score_set_test)
 
 
 def test_can_create_score_set_with_flipped_adjacent_ranges():
@@ -275,7 +270,7 @@ def test_can_create_score_set_with_flipped_adjacent_ranges():
         ],
     }
 
-    ScoreSetModify(**jsonable_encoder(score_set_test))
+    ScoreSetModify(**score_set_test)
 
 
 def test_can_create_score_set_with_adjacent_negative_ranges():
@@ -288,7 +283,7 @@ def test_can_create_score_set_with_adjacent_negative_ranges():
         ],
     }
 
-    ScoreSetModify(**jsonable_encoder(score_set_test))
+    ScoreSetModify(**score_set_test)
 
 
 def test_can_create_score_set_with_flipped_adjacent_negative_ranges():
@@ -301,7 +296,7 @@ def test_can_create_score_set_with_flipped_adjacent_negative_ranges():
         ],
     }
 
-    ScoreSetModify(**jsonable_encoder(score_set_test))
+    ScoreSetModify(**score_set_test)
 
 
 def test_cannot_create_score_set_with_overlapping_upper_unbounded_ranges():
@@ -315,7 +310,7 @@ def test_cannot_create_score_set_with_overlapping_upper_unbounded_ranges():
     }
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetModify(**jsonable_encoder(score_set_test))
+        ScoreSetModify(**score_set_test)
 
     assert "Score ranges may not overlap; `range_1` overlaps with `range_2`" in str(exc_info.value)
 
@@ -331,7 +326,7 @@ def test_cannot_create_score_set_with_overlapping_lower_unbounded_ranges():
     }
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetModify(**jsonable_encoder(score_set_test))
+        ScoreSetModify(**score_set_test)
 
     assert "Score ranges may not overlap; `range_1` overlaps with `range_2`" in str(exc_info.value)
 
@@ -347,7 +342,7 @@ def test_cannot_create_score_set_with_backwards_bounds():
     }
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetModify(**jsonable_encoder(score_set_test))
+        ScoreSetModify(**score_set_test)
 
     assert "The lower bound of the score range may not be larger than the upper bound." in str(exc_info.value)
 
@@ -362,7 +357,7 @@ def test_cannot_create_score_set_with_equal_bounds():
     }
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetModify(**jsonable_encoder(score_set_test))
+        ScoreSetModify(**score_set_test)
 
     assert "The lower and upper bound of the score range may not be the same." in str(exc_info.value)
 
@@ -378,7 +373,7 @@ def test_cannot_create_score_set_with_duplicate_range_labels():
     }
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetModify(**jsonable_encoder(score_set_test))
+        ScoreSetModify(**score_set_test)
 
     assert "Detected repeated label: `range_1`. Range labels must be unique." in str(exc_info.value)
 
@@ -394,7 +389,7 @@ def test_cannot_create_score_set_with_duplicate_range_labels_whitespace():
     }
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetModify(**jsonable_encoder(score_set_test))
+        ScoreSetModify(**score_set_test)
 
     assert "Detected repeated label: `range_1`. Range labels must be unique." in str(exc_info.value)
 
@@ -411,7 +406,7 @@ def test_cannot_create_score_set_with_wild_type_outside_ranges():
     }
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetModify(**jsonable_encoder(score_set_test))
+        ScoreSetModify(**score_set_test)
 
     assert (
         f"The provided wild type score of {wt_score} is not within any of the provided normal ranges. This score should be within a normal range."
@@ -431,7 +426,7 @@ def test_cannot_create_score_set_with_wild_type_outside_normal_range():
     }
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetModify(**jsonable_encoder(score_set_test))
+        ScoreSetModify(**score_set_test)
 
     assert (
         f"The provided wild type score of {wt_score} is not within any of the provided normal ranges. This score should be within a normal range."
@@ -450,7 +445,7 @@ def test_cannot_create_score_set_with_wild_type_score_and_no_normal_range():
     }
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetModify(**jsonable_encoder(score_set_test))
+        ScoreSetModify(**score_set_test)
 
     assert "A wild type score has been provided, but no normal classification range exists." in str(exc_info.value)
 
@@ -480,7 +475,7 @@ def test_cannot_create_score_set_without_default_ranges():
     }
 
     with pytest.raises(ValueError) as exc_info:
-        ScoreSetModify(**jsonable_encoder(score_set_test))
+        ScoreSetModify(**score_set_test)
 
     assert "unexpected value; permitted: 'normal', 'abnormal', 'not_specified'" in str(exc_info.value)
