+++ tests/worker/conftest.py
@@ -6,8 +6,8 @@ import pytest
 from mavedb.models.license import License
 from mavedb.models.taxonomy import Taxonomy
 from mavedb.models.user import User
+
 from tests.helpers.constants import EXTRA_USER, TEST_LICENSE, TEST_INACTIVE_LICENSE, TEST_TAXONOMY, TEST_USER
-from tests.helpers.util import create_experiment, create_seq_score_set
 
 
 @pytest.fixture
@@ -19,15 +19,6 @@ def setup_worker_db(session):
     db.add(License(**TEST_LICENSE))
     db.add(License(**TEST_INACTIVE_LICENSE))
     db.commit()
-
-
-@pytest.fixture
-def populate_worker_db(data_files, client):
-    # create score set via API. In production, the API would invoke this worker job
-    experiment = create_experiment(client)
-    score_set = create_seq_score_set(client, experiment["urn"])
-
-    return score_set["urn"]
 
 
 @pytest.fixture
