+++ tests/worker/test_jobs.py
@@ -1,16 +1,19 @@
+# ruff: noqa: E402
+
 from asyncio.unix_events import _UnixSelectorEventLoop
 from copy import deepcopy
 from datetime import date
 from unittest.mock import patch
 from uuid import uuid4
 
-import arq.jobs
-import cdot.hgvs.dataproviders
 import jsonschema
 import pandas as pd
 import pytest
-from arq import ArqRedis
 from sqlalchemy import not_, select
+
+arq = pytest.importorskip("arq")
+cdot = pytest.importorskip("cdot")
+fastapi = pytest.importorskip("fastapi")
 
 from mavedb.data_providers.services import VRSMap
 from mavedb.lib.mave.constants import HGVS_NT_COLUMN
@@ -31,6 +34,8 @@ from mavedb.worker.jobs import (
     map_variants_for_score_set,
     variant_mapper_manager,
 )
+
+
 from tests.helpers.constants import (
     TEST_CDOT_TRANSCRIPT,
     TEST_MINIMAL_ACC_SCORESET,
@@ -39,7 +44,18 @@ from tests.helpers.constants import (
     TEST_VARIANT_MAPPING_SCAFFOLD,
     VALID_ACCESSION,
 )
-from tests.helpers.util import awaitable_exception
+from tests.helpers.util.exceptions import awaitable_exception
+from tests.helpers.util.experiment import create_experiment
+from tests.helpers.util.score_set import create_seq_score_set
+
+
+@pytest.fixture
+def populate_worker_db(data_files, client):
+    # create score set via API. In production, the API would invoke this worker job
+    experiment = create_experiment(client)
+    score_set = create_seq_score_set(client, experiment["urn"])
+
+    return score_set["urn"]
 
 
 async def setup_records_and_files(async_client, data_files, input_score_set):
@@ -815,7 +831,7 @@ async def test_create_mapped_variants_for_scoreset_mapping_exception_retry_faile
             "run_in_executor",
             return_value=awaitable_exception(),
         ),
-        patch.object(ArqRedis, "lpush", awaitable_exception()),
+        patch.object(arq.ArqRedis, "lpush", awaitable_exception()),
     ):
         result = await map_variants_for_score_set(standalone_worker_context, uuid4().hex, score_set.id, 1)
 
@@ -909,7 +925,7 @@ async def test_create_mapped_variants_for_scoreset_parsing_exception_retry_faile
             "run_in_executor",
             return_value=dummy_mapping_job(),
         ),
-        patch.object(ArqRedis, "lpush", awaitable_exception()),
+        patch.object(arq.ArqRedis, "lpush", awaitable_exception()),
     ):
         result = await map_variants_for_score_set(standalone_worker_context, uuid4().hex, score_set.id, 1)
 
@@ -1033,7 +1049,7 @@ async def test_mapping_manager_empty_queue(setup_worker_db, standalone_worker_co
 @pytest.mark.asyncio
 async def test_mapping_manager_empty_queue_error_during_setup(setup_worker_db, standalone_worker_context):
     await standalone_worker_context["redis"].set(MAPPING_CURRENT_ID_NAME, "")
-    with patch.object(ArqRedis, "rpop", Exception()):
+    with patch.object(arq.ArqRedis, "rpop", Exception()):
         result = await variant_mapper_manager(standalone_worker_context, uuid4().hex, 1)
 
     # No new jobs should have been created if nothing is in the queue, and the queue should remain empty.
@@ -1115,7 +1131,7 @@ async def test_mapping_manager_occupied_queue_mapping_in_progress_error_during_e
     await standalone_worker_context["redis"].set(MAPPING_CURRENT_ID_NAME, "5")
     with (
         patch.object(arq.jobs.Job, "status", return_value=arq.jobs.JobStatus.in_progress),
-        patch.object(ArqRedis, "enqueue_job", return_value=awaitable_exception()),
+        patch.object(arq.ArqRedis, "enqueue_job", return_value=awaitable_exception()),
     ):
         result = await variant_mapper_manager(standalone_worker_context, uuid4().hex, 1)
 
@@ -1143,7 +1159,7 @@ async def test_mapping_manager_occupied_queue_mapping_not_in_progress_error_duri
     await standalone_worker_context["redis"].set(MAPPING_CURRENT_ID_NAME, "")
     with (
         patch.object(arq.jobs.Job, "status", return_value=arq.jobs.JobStatus.not_found),
-        patch.object(ArqRedis, "enqueue_job", return_value=awaitable_exception()),
+        patch.object(arq.ArqRedis, "enqueue_job", return_value=awaitable_exception()),
     ):
         result = await variant_mapper_manager(standalone_worker_context, uuid4().hex, 1)
